### 2024-07-08 156f60e
Update 'spectrogram_images' function to avoid data leakage.
