+++ R/evaluate_trainedmodel_performance_multi.R
@@ -15,36 +15,32 @@
 # if (.Platform$OS.type == "windows") {
 #   # Set directory paths for trained models and test images
 #
-#   trained_models_dir <- system.file("extdata", "trainedresnetbinary\\", package = "gibbonNetR")
-#   image_data_dir <- system.file("extdata", "binary\\test\\", package = "gibbonNetR")
-#
+#   trained_models_dir <- system.file("extdata", "trainedresnetmulti\\", package = "gibbonNetR")
+#   image_data_dir <- system.file("extdata", "multiclass\\test\\", package = "gibbonNetR")
+#   class_names <-  c('female.gibbon','hornbill.helmeted','hornbill.rhino','long.argus','noise')
 #   # Evaluate the performance of the trained models using the test images
-#   evaluate_trainedmodel_performance(trained_models_dir = trained_models_dir,
-#                                     image_data_dir = image_data_dir,
-#                                     output_dir = paste(tempdir(), '\\data\\',sep=''),
-#                                     positive.class = 'Gibbons',  #' Label for positive class
-#                                     negative.class = 'Noise')    #' Label for negative class
-#
+#   evaluate_trainedmodel_performance_multi(trained_models_dir=trained_models_dir,
+#                                           class_names=class_names,
+#                                           image_data_dir=image_data_dir,
+#                                           output_dir= paste(tempdir(), '\\data\\',sep=''),
+#                                           noise.category = "noise")
 #   # Find the location of saved evaluation files
 #   CSVName <- list.files(paste(tempdir(), '\\data\\',sep=''), recursive = TRUE, full.names = TRUE)
 #
 #   # Check the output of the first file
 #   head(read.csv(CSVName[1]))
 #
-# } else  {
+# } else {
 #
-#   # Set directory paths for trained models and test images
-#
-#   trained_models_dir <- system.file("extdata", "trainedresnetbinary/", package = "gibbonNetR")
-#   image_data_dir <- system.file("extdata", "binary/test/", package = "gibbonNetR")
-#
+#   trained_models_dir <- system.file("extdata", "trainedresnetmulti/", package = "gibbonNetR")
+#   image_data_dir <- system.file("extdata", "multiclass/test/", package = "gibbonNetR")
+#   class_names <-  c('female.gibbon','hornbill.helmeted','hornbill.rhino','long.argus','noise')
 #   # Evaluate the performance of the trained models using the test images
-#   evaluate_trainedmodel_performance(trained_models_dir = trained_models_dir,
-#                                     image_data_dir = image_data_dir,
-#                                     output_dir = paste(tempdir(), '/data/'),
-#                                     positive.class = 'Gibbons',  #' Label for positive class
-#                                     negative.class = 'Noise')    #' Label for negative class
-#
+#   evaluate_trainedmodel_performance_multi(trained_models_dir=trained_models_dir,
+#                                           class_names=class_names,
+#                                           image_data_dir=image_data_dir,
+#                                           output_dir= paste(tempdir(), '/data/'),
+#                                           noise.category = "noise")
 #   # Find the location of saved evaluation files
 #   CSVName <- list.files(paste(tempdir(), '/data/'), recursive = TRUE, full.names = TRUE)
 #
@@ -53,6 +49,7 @@
 #
 # }
 #'}
+#'
 #' @importFrom purrr %>%
 #' @export
 #'
