### 2024-08-07 #6801 @DenaJGibbon
Sorry I missed the earlier messages - will make sure to check back here more often. After a quick scan through their GitHub pages I do not think there are any COIs.

I also wonder if any of the authors from this R package might work? It is not acoustics, but uses torch for R, and
https://www.biorxiv.org/content/10.1101/2022.02.07.479461v1
