### 2024-08-25 #6801 @faroit
@DenaJGibbon yes, of course. You can update the paper at any time. Just use the same branch so that the automatic build process can find the right paper. Also the review process is fluid so you can also update the paper during that phase as well.
