### 2024-08-26 #6801 @Desjonqu
Just to be sure, I will make my assessment and submit it here, is that right?
