### 2024-08-26 #6801 @steffilazerte
Hi @faroit, I can review with respect to R and R packages and superficially acoustics (my background is a bit dated and more targeted song analysis, no broad bioacoustics and no specialized software), but it might be better to get someone with better expertise than me.

Have you considered [Marcelo Araya-Salas](https://github.com/maRce10)? Possibly also [Jerome Sueur](https://github.com/jeromesueur) (author of [Seewave](https://rug.mnhn.fr/seewave/))?
