### 2024-09-20 #7250 @editorialbot
[REVIEW]: gibbonNetR: an R Package for the Use of Convolutional Neural Networks and TransferLearning on Acoustic Data
**Submitting author:** @DenaJGibbon (<a href="http://orcid.org/0000-0003-0363-5581">Dena Clink</a>)
**Repository:** https://github.com/DenaJGibbon/gibbonNetR
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
