### 2024-10-01 DenaJGibbon/gibbonNetR#2 @faroit
[JOSS] fix invalid DOIs
@DenaJGibbon please update the paper reference so that DOIs are used for all papers if possible

```
❌ MISSING DOIs

- 10.3389/fevo.2023.1071640 may be a valid DOI for title: A workflow for the automated detection and classif...
- 10.1080/09524622.2015.1133320 may be a valid DOI for title: Assessment of error rates in acoustic monitoring w...
- 10.1109/cvpr.2009.5206848 may be a valid DOI for title: Imagenet: A large-scale hierarchical image databas...
- 10.1201/9781003275923 may be a valid DOI for title: Deep Learning and Scientific Computing with R torc...
- 10.1145/3065386 may be a valid DOI for title: Imagenet classification with deep convolutional ne...
- 10.1109/cvpr.2016.90 may be a valid DOI for title: Deep residual learning for image recognition
- 10.1016/j.ecoinf.2023.102457 may be a valid DOI for title: Mel-frequency cepstral coefficients outperform emb...
- 10.1038/s41598-023-49989-z may be a valid DOI for title: Global birdsong embeddings enable superior transfe...
- 10.1371/journal.pone.0246564 may be a valid DOI for title: Not by the light of the moon: investigating circad...
- 10.1111/btp.13205 may be a valid DOI for title: Evidence for acoustic niche partitioning depends o...

❌ INVALID DOIs

- https://doi.org/10.1016/j.ecoinf.2022.101688 is INVALID because of 'https://doi.org/' prefix
- https://doi.org/10.1016/j.apacoust.2022.108939 is INVALID because of 'https://doi.org/' prefix
- https://doi.org/10.1016/j.ecoinf.2023.102457 is INVALID because of 'https://doi.org/' prefix
```

https://github.com/openjournals/joss-reviews/issues/7250
