### 2024-10-01 DenaJGibbon/gibbonNetR#3 @faroit
[JOSS] move to one license file
i think JOSS doesn't find the license.md file so please move over the license text to [`LICENSE`](https://github.com/DenaJGibbon/gibbonNetR/blob/main/LICENSE)

https://github.com/openjournals/joss-reviews/issues/7250
