### 2024-10-23 DenaJGibbon/gibbonNetR#4 @steffilazerte
[JOSS review] pkgdown documentation site
[Hi! I'm Steffi, one of your JOSS reviewers. I'm opening issues in your repository to hold the details of my comments for the review. I'll let you know when I'm done in the [main review thread](https://github.com/openjournals/joss-reviews/issues/7250)]

Consider adding a pkgdown documentation website (see [pkgdown](https://pkgdown.r-lib.org/)).

Although it looks complex, it's a highly automated process and can be up and [running in only a few minutes](https://pkgdown.r-lib.org/#usage). This is a great place for users to see at a glance all the functions available (you can even organize the function reference to make it easier for users to understand them). Compiling this website also helps you find and fix any errors that may pop up in a different environment than then one you typically work in.

Examples in your function docs will also automatically compile so users can see what the outputs are supposed to look like!

I don't think this is required for JOSS, but as a way of enhancing the documentation and visibility of your package, I highly recommend it!

CC @faroit
