### 2024-10-23 DenaJGibbon/gibbonNetR#5 @steffilazerte
[JOSS review] Documentation
[Hi! I'm Steffi, one of your JOSS reviewers. I'm opening issues in your repository to hold the details of my comments for the review. I'll let you know when I'm done in the [main review thread](https://github.com/openjournals/joss-reviews/issues/7250)]

- [x] Consider including package-level documentation (e.g., https://r-pkgs.org/man.html#sec-man-package-doc).
- [x] I love that you have examples for nearly all the exported functions, that's a great way to help the user understand what's going on. Many of the examples seem a bit complicated by the `if()/else()` for platform type. However, I don't think you should need this for `system.file()` and if you use `file.path()`  instead of `paste()`ing the paths together. Then you can remove the `if()/else()` statements and clarify those examples.
- [x] In your function documentation, if one function is likely to follow another, or if the inputs of one are the outputs of another, add these details to the function description and argument descriptions. That will help users understand the workflow.
- [x] Consider adding some more descriptions of your included example data to make it easier for users to understand what the important bits are and how they should prepare their own data. This could be in a vignette or the README.
- [x] I was a bit confused about the workflow and order of operations. I'm not sure what functions should be used when, or what kind of data is required as an input. You could help clarify this by adding explanations to the README examples (see issue README), but I would also consider having a broad overview of when to use what function in an vignette/article (see #4). If you decide against the pkgdown site, you could also add this explanation to the package-level documentation.
- [x] If you believe that people using this package should have an understanding of machine learning, then I would recommend including in your package Documentation/README a statement to this effect and links to where people can get the foundation they require before using your package.

CC @faroit
