### 2024-10-23 DenaJGibbon/gibbonNetR#6 @steffilazerte
[JOSS review] README
[Hi! I'm Steffi, one of your JOSS reviewers. I'm opening issues in your repository to hold the details of my comments for the review. I'll let you know when I'm done in the [main review thread](https://github.com/openjournals/joss-reviews/issues/7250)]

- [x] It would be helpful to include a brief statement of what this package does and why someone might want to use it right at the top of the README.
- [x] It's really nice that the README has so many examples, and I think a bit more context would really help the user understand how to get started with their own data. At the very least, I think it could really help the user if you gave a brief description of what each example does. Why would you run that code? What are the inputs? How would you interpret the outputs? If you want to take it to the next level, I suggest moving each example to it's own vignette/article (see #4), and adding additional details. Treat it like a tutorial where you are guiding a colleague by the hand.

A couple of specific problems I ran into which made it hard for me to evaluate whether the functions work as they're supposed to:

- [x] The 'download example test files' is presented twice in the README
- [x] In the download example, you need to create the "data" directory first, consider adding `dir.create("data")`  before the `download.file()` step.
- [x] Running `spectrogram_images()` in the download example I got a lot of messages like "data/trainingimages//valid/female.gibbon already exists saving spectrogram images", "data/trainingimages//train/female.gibbon already exists saving spectrogram images", is this expected? (also for noise). After all the messages, this function then ended with an `Error in tuneR::readWave(SoundFiles[y]) : File 'NA' does not exist.`
- [x] In the Train the models examples, I got an `error: "Error in (function (self, target, weight, reduction, ignore_index, label_smoothing)  :weight tensor should be defined either for all 6 classes or no classes but got weight tensor of shape"` (and more) - Possibly this is because of the earlier problems? Or are these examples unrelated to each other?

On the bright side, I was able to run all your testthat tests no problem, so perhaps it's a problem with the way the examples are setup?

CC @faroit
