### 2024-10-23 DenaJGibbon/gibbonNetR#7 @steffilazerte
[JOSS review] General R package best practices and fixing errors
[Hi! I'm Steffi, one of your JOSS reviewers. I'm opening issues in your repository to hold the details of my comments for the review. I'll let you know when I'm done in the [main review thread](https://github.com/openjournals/joss-reviews/issues/7250)]

- [ ] I'm used to the [rOpenSci review process](https://ropensci.org/software-review/) which is a bit more firm about adoption of best practices. I understand that that's not strictly a requirement for the JOSS review, but some best practices are important to make your package more reliable and less fragile which can make your life easier as well as make for a better user experience.
- [ ] Consider running [pkgcheck](https://docs.ropensci.org/pkgcheck/) and reviewing it's suggestions. For example, things like using TRUE vs. T are good to fix because they can introduce problematic errors otherwise.
- [ ] Try to use `message()` rather than `print()` , as that allows users to suppress the progress messages if they want.
- [ ] There are quite a few packages in "Depends", some may make sense if you use their functions a lot, but generally if you can reference the functions like `ggplot2::ggplot(data = x, ggplot2::aes(x = ..., y =...))` it's safer and nicer to your user (see the R Packages guide for more specifics: https://r-pkgs.org/dependencies-in-practice.html)
- [ ] Run package [check](https://r-pkgs.org/whole-game.html#check) (In RStudio with devtools, this can be run with Ctrl-Shift-E
    - [ ] Fix as many notes as you can (i.e. use `utils::read.csv()` instead of just `read.csv()` to fix the "no visible global function definition for 'read.csv')
    - [ ] Fix the warning
    - [ ] Remove `devtools::load_all("/Users/denaclink/Desktop/RStudioProjects/gibbonNetR")` from testthat.R to fix running the tests
    - [ ] Fix the errors in the examples

Feel free to let me know if you have any questions about how to fix any of these warnings/errors etc.!

CC @faroit
