### 2024-10-23 DenaJGibbon/gibbonNetR#8 @steffilazerte
[JOSS review] Manuscript
[Hi! I'm Steffi, one of your JOSS reviewers. I'm opening issues in your repository to hold the details of my comments for the review. I'll let you know when I'm done in the [main review thread](https://github.com/openjournals/joss-reviews/issues/7250)]

I really appreciated the overview at the start of the paper and the first paragraph of the Statement of need. They gave me a clear idea of what the purpose of this software is and what role it serves. I was impressed and pretty excited about the software! Later, I did get a bit bogged down in the following paragraphs on Automated Detection and Transfer Learning. It wasn't clear to me how these processes worked. Machine learning is definitely not my background, but that might be the case with many of your users. I think it could be worth thinking of your audience as enthusiastic (behavioural?) ecologists, with R skills, but not really that well versed in Machine Learning. What is important about this software for them to know? How would you explain the nuances among these approaches to them?

I'm also not sure the Usage section is required. Perhaps consider showing only the basic usage and moving the rest to a vignette/article in the package documentation. Better to have good documentation in your package and link to that. The usage of your package may change over time, but the premise behind it, probably less so.  @faroit, do these recommendations make sense for a JOSS article?

With respect to the State of the Field, I wonder if it might be more useful to consider the software in the context of acoustic software, rather than Python vs. R. What other acoustic R packages are there for doing this kind of work? Or what Python packages are there? Are there specific acoustic packages? How does gibbonNetR compare these? Did you want to mention your other package, gibbonR here?

Finally, in the Future Directions, there's no need to say that your methods aren't state of the art. Do you really think the only value your package has is for benchmarking? If you consider what other software is available in R acoustics, does that change your opinion? I think you should be able to give a more encouraging statement of how useful your package can be!

### Nitty gritty

- [x] Automated Detection - Why is it a problem to rely on feature engineering?
- [x] Line 75, duplicated sentence?
- [x] LIne 81, examples? Or do you mean that your tutorials/examples are borrowed from theirs?
- [x] Line 84, "which are used for transfer learning"? Should you mention that the spectrograms are from seewave?
- [x] L90 Which species?
- [x] L93 classes?
- [x] Figures hav double fig reference
- [x] If you move examples to a vignette/article, I would also add some details to your examples with respect to why would you want to (i.e.) Plot the unsupervised clustering results? In the section Extract embeddings you said why we might want to which is great, and I would also suggest explaining what you would do with this extracted results.

CC @faroit
