### 2024-10-23 DenaJGibbon/gibbonNetR#9 @steffilazerte
[JOSS review] R package check
[Hi! I'm Steffi, one of your JOSS reviewers. I'm opening issues in your repository to hold the details of my comments for the review. I'll let you know when I'm done in the [main review thread](https://github.com/openjournals/joss-reviews/issues/7250)]

Run package [check](https://r-pkgs.org/whole-game.html#check) (In RStudio with devtools, this can be run with Ctrl-Shift-E

- [x] Fix as many notes as you can (i.e. use `utils::read.csv()` instead of just `read.csv()` to fix the "no visible global function definition for 'read.csv')
- [x] Fix the warning
- [x] Remove `devtools::load_all("/Users/denaclink/Desktop/RStudioProjects/gibbonNetR")` from testthat.R to fix running the tests
- [x] Fix the errors in the examples

Feel free to let me know if you have any questions about how to fix any of these warnings/errors etc.!

CC @faroit
