### 2024-10-28 #7250 @DenaJGibbon
Thank you so much for your detailed comments, @steffilazerte ! They will be very helpful for improving the package and the manuscript. We will get going on the revisions!
