+++ paper/references.bib
@@ -287,7 +287,7 @@ Publisher: Nature Publishing Group},
 
 @article{lakdari2024mel,
   title={Mel-frequency cepstral coefficients outperform embeddings from pre-trained convolutional neural networks under noisy conditions for discrimination tasks of individual gibbons},
-  author={Lakdari, Mohamed Walid and Ahmad, Abdul Hamid and Sethi, Sarab and Bohn, Gabriel A and Clink, Dena J},
+  author={Lakdari, Mohamed Walid and Ahmad, Abdul Hamid and Sethi, Sarab and Bohn, Gabriel A and Clink, Dena J.},
   journal={Ecological Informatics},
   volume={80},
   pages={102457},
@@ -297,7 +297,7 @@ Publisher: Nature Publishing Group},
 
 @article{sethi2020characterizing,
   title={Characterizing soundscapes across diverse ecosystems using a universal acoustic feature set},
-  author={Sethi, Sarab S and Jones, Nick S and Fulcher, Ben D and Picinali, Lorenzo and Clink, Dena Jane and Klinck, Holger and Orme, C David L and Wrege, Peter H and Ewers, Robert M},
+  author={Sethi, Sarab S and Jones, Nick S and Fulcher, Ben D and Picinali, Lorenzo and Clink, Dena J. and Klinck, Holger and Orme, C David L and Wrege, Peter H and Ewers, Robert M},
   journal={Proceedings of the National Academy of Sciences},
   volume={117},
   number={29},
@@ -330,7 +330,7 @@ Publisher: Nature Publishing Group},
 }
 @article{kennedy2023evidence,
   title={Evidence for acoustic niche partitioning depends on the temporal scale in two sympatric Bornean hornbill species},
-  author={Kennedy, Amy G and Ahmad, Abdul Hamid and Klinck, Holger and Johnson, Lynn M and Clink, Dena J},
+  author={Kennedy, Amy G and Ahmad, Abdul Hamid and Klinck, Holger and Johnson, Lynn M and Clink, Dena J.},
   journal={Biotropica},
   volume={55},
   number={2},
@@ -355,7 +355,7 @@ Publisher: Nature Publishing Group},
 }
 @article{10.1371/journal.pone.0283396,
     doi = {10.1371/journal.pone.0283396},
-    author = {Best, Paul AND Paris, Sébastien AND Glotin, Hervé AND Marxer, Ricard},
+    author = {Best, Paul and Paris, Sébastien and Glotin, Hervé and Marxer, Ricard},
     journal = {PLOS ONE},
     publisher = {Public Library of Science},
     title = {Deep audio embeddings for vocalisation clustering},
@@ -377,7 +377,7 @@ year = {2024},
 issn = {1574-9541},
 doi = {https://doi.org/10.1016/j.ecoinf.2023.102457},
 url = {https://www.sciencedirect.com/science/article/pii/S1574954123004867},
-author = {Mohamed Walid Lakdari and Abdul Hamid Ahmad and Sarab Sethi and Gabriel A. Bohn and Dena J. Clink},
+author = {Lakdari,Mohamed Walid  and Ahmad,Abdul Hamid and Sethi, Sarab and Bohn, Gabriel A.  and Clink, Dena J.},
 keywords = {Vocal individuality, Sound feature extraction, Mel-frequency cepstral coefficients, Convolutional neural networks, Acoustic indices, },
[…]
