+++ vignettes/gibbonNetR.Rmd
@@ -1,0 +1,19 @@
+---
+title: "gibbonNetR"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{gibbonNetR}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+```{r setup}
+library(gibbonNetR)
+```
