+++ DESCRIPTION
@@ -45,3 +45,4 @@ Suggests:
 Config/testthat/edition: 3
 VignetteBuilder: knitr
 LazyData: true
+URL: https://denajgibbon.github.io/gibbonNetR/
