+++ _pkgdown.yml
@@ -1,0 +1,4 @@
+url: https://denajgibbon.github.io/gibbonNetR/
+template:
+  bootstrap: 5
+
