+++ pkgdown.yml
@@ -1,0 +1,9 @@
+pandoc: 3.1.11
+pkgdown: 2.1.1
+pkgdown_sha: ~
+articles:
+  gibbonNetR: gibbonNetR.html
+last_built: 2024-11-23T16:01Z
+urls:
+  reference: https://denajgibbon.github.io/gibbonNetR/reference
+  article: https://denajgibbon.github.io/gibbonNetR/articles
