### 2024-11-23 7115c66
First vignette commit.
