+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
   gibbonNetR: gibbonNetR.html
-last_built: 2024-11-23T16:01Z
+last_built: 2024-11-23T16:06Z
 urls:
   reference: https://denajgibbon.github.io/gibbonNetR/reference
   article: https://denajgibbon.github.io/gibbonNetR/articles
