+++ vignettes/gibbonNetR.Rmd
@@ -1,6 +1,10 @@
 ---
 title: "gibbonNetR"
-output: rmarkdown::html_vignette
+output:
+  rmarkdown::html_vignette:
+    toc: true
+    toc_depth: 1
+    number_sections: true
 vignette: >
   %\VignetteIndexEntry{gibbonNetR}
   %\VignetteEngine{knitr::rmarkdown}
@@ -14,7 +18,76 @@ knitr::opts_chunk$set(
 )
 ```
 
-First load the packages
-```{r setup, message=FALSE}
+# Data Preparation
+
+## First load the packages
+
+```{r message=FALSE}
 library(gibbonNetR)
+```
+
+## We create spectrogram images of the sound files. 
+To follow best practices we want to use training and test data from different recording locations. These clips are from Danum Valley Conservation Area, Sabah, Malaysia, but are taken from two different passive acoustic monitoring arrays. The "spectrogram_images" takes 
+
+```{r eval=FALSE}
+# Link to training clips on Zenodo
+ZenodoLink <- 'https://zenodo.org/records/11355157/files/trainingclips.zip?download=1'
+
+# Download into specified zip file location
+download.file(url = ZenodoLink, destfile = 'data/spectrogram_image_examples.zip',method='curl')
+
+# Unzip folder
+exdir <- 'data/'
+utils::unzip(zipfile = 'data/spectrogram_image_examples.zip', exdir = exdir )
+
+# Check folder composition
+TrainingDatapath <- paste(exdir,"trainingclips",sep='')
+
+# Check folder names
+list.files(TrainingDatapath)
+
+# Create spectrogram images
+spectrogram_images(
+   trainingBasePath = TrainingDatapath,
+   outputBasePath = 'data/trainingimages/',
+   minfreq.khz = 0.4,
+   maxfreq.khz = 1.6,
+   splits = c(0.7, 0.3, 0), # Assign proportion to training, validation, or test folders
+   new.sampleratehz = 'NA'
+ )
+
+
+```
+
+## Download example test files on Zenodo and convert to spectrogram images
+
+```{r eval = FALSE}
+library(gibbonNetR)
+
+# Link to training clips on Zenodo
+ZenodoLink <- 'https://zenodo.org/records/11355157/files/testclips.zip?download=1'
+
+# Download into specified zip file location
+download.file(url = ZenodoLink, destfile = 'data/spectrogram_image_examples.zip',method='curl')
+
+# Unzip folder
+exdir <- 'data/'
+utils::unzip(zipfile = 'data/spectrogram_image_examples.zip', exdir = exdir )
+
+# Check folder composition
+TestDatapath <- paste(exdir,"testclips",sep='')
+
+# Check folder names
+list.files(TestDatapath)
+
+# Create spectorgram images
+spectrogram_images(
+   trainingBasePath = TestDatapath,
+   outputBasePath = 'data/testimages/',
+   minfreq.khz = 0.4,
+   maxfreq.khz = 1.6,
+   splits = c(0, 0, 1), # Assign proportion to training, validation, or test folders
+   new.sampleratehz = 'NA'
+ )
+
 ```
