+++ vignettes/gibbonNetR.Rmd
@@ -27,18 +27,19 @@ library(gibbonNetR)
 ```
 
 ## We create spectrogram images of the sound files. 
-To follow best practices we want to use training and test data from different recording locations. These clips are from Danum Valley Conservation Area, Sabah, Malaysia, but are taken from two different passive acoustic monitoring arrays. The "spectrogram_images" takes 
+To follow best practices we want to use training and test data from different recording locations. These clips are from Danum Valley Conservation Area, Sabah, Malaysia, but are taken from two different passive acoustic monitoring arrays. The "spectrogram_images" function takes folders organized and labeled by signal type, and creates training, validation, or test folders using the "splits" specified in the function call.   
 
 ```{r eval=FALSE}
 # Link to training clips on Zenodo
-ZenodoLink <- 'https://zenodo.org/records/11355157/files/trainingclips.zip?download=1'
+ZenodoLink <- 'https://zenodo.org/records/14210125/files/trainingclips.zip?download=1'
+
 
 # Download into specified zip file location
-download.file(url = ZenodoLink, destfile = 'data/spectrogram_image_examples.zip',method='curl')
+download.file(url = ZenodoLink, destfile = 'data/trainingclips.zip',method='curl')
 
 # Unzip folder
 exdir <- 'data/'
-utils::unzip(zipfile = 'data/spectrogram_image_examples.zip', exdir = exdir )
+utils::unzip(zipfile = 'data/trainingclips.zip', exdir = exdir )
 
 # Check folder composition
 TrainingDatapath <- paste(exdir,"trainingclips",sep='')
@@ -52,6 +53,7 @@ spectrogram_images(
    outputBasePath = 'data/trainingimages/',
    minfreq.khz = 0.4,
    maxfreq.khz = 1.6,
+      random=FALSE,
    splits = c(0.7, 0.3, 0), # Assign proportion to training, validation, or test folders
    new.sampleratehz = 'NA'
  )
@@ -65,14 +67,14 @@ spectrogram_images(
 library(gibbonNetR)
 
 # Link to training clips on Zenodo
-ZenodoLink <- 'https://zenodo.org/records/11355157/files/testclips.zip?download=1'
+ZenodoLink <- 'https://zenodo.org/records/14210125/files/testclips.zip?download=1'
 
 # Download into specified zip file location
-download.file(url = ZenodoLink, destfile = 'data/spectrogram_image_examples.zip',method='curl')
+download.file(url = ZenodoLink, destfile = 'data/testclips.zip',method='curl')
 
 # Unzip folder
 exdir <- 'data/'
-utils::unzip(zipfile = 'data/spectrogram_image_examples.zip', exdir = exdir )
+utils::unzip(zipfile = 'data/testclips.zip', exdir = exdir )
 
 # Check folder composition
 TestDatapath <- paste(exdir,"testclips",sep='')
@@ -91,3 +93,65 @@ spectrogram_images(
  )
 
 ```
+
+# Model training
+```{r eval = FALSE}
+# Location of spectrogram images for training
+input.data.path <-  'data/trainingimages/'
+
+# Location of spectrogram images for testing
+test.data.path <- 'data/testimages/test/'
+
+# User specified training data label for metadata
+trainingfolder.short <- 'danummulticlassexample'
+
+# Specify the architecture type
+architecture <- c('alexnet','resnet50')
+
+# We can specify the number of epochs to train here
+epoch.iterations <- c(1)
+
+# Function to train a multi-class CNN
+gibbonNetR::train_CNN_multi(input.data.path=input.data.path,
+                            architecture =architecture,
+                            learning_rate = 0.001,
+                            class_weights = c(0.3, 0.3, 0.2, 0.2, 0),
+                            test.data=test.data.path,
+                            unfreeze.param = TRUE,
[…]
