+++ README.Rmd
@@ -79,14 +79,14 @@ library(purrr)
 library(gibbonNetR)
 
 # Link to training clips on Zenodo
-ZenodoLink <- 'https://zenodo.org/records/11355157/files/trainingclips.zip?download=1'
+ZenodoLink <- 'https://zenodo.org/records/14210125/files/trainingclips.zip?download=1'
 
 # Download into specified zip file location
-download.file(url = ZenodoLink, destfile = 'data/spectrogram_image_examples.zip',method='curl')
+download.file(url = ZenodoLink, destfile = 'data/trainingclips.zip',method='curl')
 
 # Unzip folder
 exdir <- 'data/'
-utils::unzip(zipfile = 'data/spectrogram_image_examples.zip', exdir = exdir )
+utils::unzip(zipfile = 'data/trainingclips.zip', exdir = exdir )
 
 # Check folder composition
 TrainingDatapath <- paste(exdir,"trainingclips",sep='')
@@ -100,6 +100,7 @@ spectrogram_images(
    outputBasePath = 'data/trainingimages/',
    minfreq.khz = 0.4,
    maxfreq.khz = 1.6,
+   random=FALSE,
    splits = c(0.7, 0.3, 0), # Assign proportion to training, validation, or test folders
    new.sampleratehz = 'NA'
  )
@@ -119,14 +120,14 @@ knitr::include_graphics("/Users/denaclink/Desktop/RStudioProjects/gibbonNetR/REA
 library(gibbonNetR)
 
 # Link to training clips on Zenodo
-ZenodoLink <- 'https://zenodo.org/records/11355157/files/testclips.zip?download=1'
+ZenodoLink <- 'https://zenodo.org/records/14210125/files/testclips.zip?download=1'
 
 # Download into specified zip file location
-download.file(url = ZenodoLink, destfile = 'data/spectrogram_image_examples.zip',method='curl')
+download.file(url = ZenodoLink, destfile = 'data/testclips.zip',method='curl')
 
 # Unzip folder
 exdir <- 'data/'
-utils::unzip(zipfile = 'data/spectrogram_image_examples.zip', exdir = exdir )
+utils::unzip(zipfile = 'data/testclips.zip', exdir = exdir )
 
 # Check folder composition
 TestDatapath <- paste(exdir,"testclips",sep='')
@@ -143,6 +144,7 @@ spectrogram_images(
    splits = c(0, 0, 1), # Assign proportion to training, validation, or test folders
    new.sampleratehz = 'NA'
  )
+
 
 ```
 # Train the models
