+++ vignettes/gibbonNetR.Rmd
@@ -149,16 +149,13 @@ PerformanceOutput <- gibbonNetR::get_best_performance(performancetables.dir=perf
 
 ```
 
-```{r, echo=FALSE}
-knitr::include_graphics("/Users/denaclink/Desktop/RStudioProjects/gibbonNetR/images/multiclass_eval.png")
-```
-
 !["Sample plot from 'get_best_performance' function "\label{fig:fig1}](/Users/denaclink/Desktop/RStudioProjects/gibbonNetR/images/multiclass_eval.png)
 ## Examine the results
 
 ```{r, eval=FALSE}
 PerformanceOutput$f1_plot
 PerformanceOutput$best_f1$F1
+[1] 0.4862252
 ```
 
 
