+++ vignettes/gibbonNetR.Rmd
@@ -5,17 +5,27 @@ output:
     toc: true
     toc_depth: 1
     number_sections: true
+    self_contained: yes
 vignette: >
   %\VignetteIndexEntry{gibbonNetR}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
 
-```{r, include = FALSE}
+```{r, eval = TRUE, echo = FALSE, message=FALSE}
+# global option chunks
 knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
+  fig.width = 5, 
+  fig.height = 3.5,
+  dpi = 70,
+  comment = "",
+  out.width = "80%",
+  fig.align = "center",
+  message = TRUE,
+  warning = TRUE
 )
+
+options(width = 100, max.print = 100)
 ```
 
 # Data Preparation
@@ -139,7 +149,7 @@ PerformanceOutput <- gibbonNetR::get_best_performance(performancetables.dir=perf
 
 ```
 
-```{r,echo=FALSE, eval=FALSE}
+```{r,echo=FALSE, eval=TRUE}
 # Evaluate model performance
 performancetables.dir <- "/Users/denaclink/Desktop/RStudioProjects/gibbonNetR/data/model_output/_danummulticlassexample_multi_unfrozen_TRUE_/performance_tables_multi"
 
@@ -149,13 +159,14 @@ PerformanceOutput <- gibbonNetR::get_best_performance(performancetables.dir=perf
 
 ```
 
-!["Sample plot from 'get_best_performance' function "\label{fig:fig1}](/Users/denaclink/Desktop/RStudioProjects/gibbonNetR/images/multiclass_eval.png)
+<!-- !["Sample plot from 'get_best_performance' function "\label{fig:fig1}](/Users/denaclink/Desktop/RStudioProjects/gibbonNetR/images/multiclass_eval.png) -->
+
 ## Examine the results
 
-```{r, eval=FALSE}
+```{r, eval=TRUE}
 PerformanceOutput$f1_plot
 PerformanceOutput$best_f1$F1
-[1] 0.4862252
+#[1] 0.4862252
 ```
 
 
