+++ vignettes/gibbonNetR.Rmd
@@ -149,7 +149,7 @@ PerformanceOutput <- gibbonNetR::get_best_performance(performancetables.dir=perf
 
 ```
 
-```{r,echo=FALSE, eval=TRUE}
+```{r,eval=FALSE}
 # Evaluate model performance
 performancetables.dir <- "/Users/denaclink/Desktop/RStudioProjects/gibbonNetR/data/model_output/_danummulticlassexample_multi_unfrozen_TRUE_/performance_tables_multi"
 
@@ -159,14 +159,15 @@ PerformanceOutput <- gibbonNetR::get_best_performance(performancetables.dir=perf
 
 ```
 
-<!-- !["Sample plot from 'get_best_performance' function "\label{fig:fig1}](/Users/denaclink/Desktop/RStudioProjects/gibbonNetR/images/multiclass_eval.png) -->
+
+<img src="multiclass_eval.png">
 
 ## Examine the results
 
-```{r, eval=TRUE}
+```{r, eval=FALSE}
 PerformanceOutput$f1_plot
 PerformanceOutput$best_f1$F1
-#[1] 0.4862252
+[1] 0.4862252
 ```
 
 
