### 2024-11-24 e781bf7
Update vignette image
