+++ vignettes/gibbonNetR.Rmd
@@ -160,15 +160,25 @@ PerformanceOutput <- gibbonNetR::get_best_performance(performancetables.dir=perf
 ```
 
 
-<img src="multiclass_eval.png">
-
 ## Examine the results
 
 ```{r, eval=FALSE}
 PerformanceOutput$f1_plot
+
 PerformanceOutput$best_f1$F1
 [1] 0.4862252
+
+PerformanceOutput$best_auc$AUC
+[1] 0.2858019
+
+PerformanceOutput$best_precision$Precision
+[1] 0.321817
+
+PerformanceOutput$best_recall$Recall
+[1] 1
 ```
 
 
-
+```{r, echo=FALSE, out.width="50%", fig.cap="Output from 'get_best_performance' fuction ."}
+knitr::include_graphics("multiclass_eval.png")
+```
