### 2024-11-24 d104bd2
Update spectro image
