+++ README.md
@@ -6,9 +6,14 @@ Dena J. Clink and Abdul Hamid Ahmad
 
 # Overview
 
-This README provides instructions and code for training and testing the
-performance of different convolutional neural network model
-architectures on spectrogram images.
+This README provides code for training and testing the performance of
+different convolutional neural network model architectures on
+spectrogram images.
+
+# Usage
+
+A detailed usage guide can be found at:
+denajgibbon.github.io/gibbonNetR/
 
 # Installation
 
@@ -51,234 +56,3 @@ library(torch)
     noise.category = 'noise'
   )
 ```
-
-# Preparing the data
-
-## Download example training files on Zenodo and convert to spectrogram images
-
-``` r
-library(gibbonNetR)
-
-# Link to training clips on Zenodo
-ZenodoLink <- 'https://zenodo.org/records/14210125/files/trainingclips.zip?download=1'
-
-# Download into specified zip file location
-download.file(url = ZenodoLink, destfile = 'data/trainingclips.zip',method='curl')
-
-# Unzip folder
-exdir <- 'data/'
-utils::unzip(zipfile = 'data/trainingclips.zip', exdir = exdir )
-
-# Check folder composition
-TrainingDatapath <- paste(exdir,"trainingclips",sep='')
-
-# Check folder names
-list.files(TrainingDatapath)
-
-# Create spectrogram images
-spectrogram_images(
-   trainingBasePath = TrainingDatapath,
-   outputBasePath = 'data/trainingimages/',
-   minfreq.khz = 0.4,
-   maxfreq.khz = 1.6,
-   random=FALSE,
-   splits = c(0.7, 0.3, 0), # Assign proportion to training, validation, or test folders
-   new.sampleratehz = 'NA'
- )
-```
-
-## Here are a few spectrogram images
-
-<div class="figure">
-
-<img src="README_files/spectro.png" alt="Figure 1. Spectrograms of training clips for CNNs" width="400px" />
-<p class="caption">
-Figure 1. Spectrograms of training clips for CNNs
-</p>
-
-</div>
-
-## Download example test files on Zenodo and convert to spectrogram images
-
-``` r
-library(gibbonNetR)
-
-# Link to training clips on Zenodo
-ZenodoLink <- 'https://zenodo.org/records/14210125/files/testclips.zip?download=1'
-
-# Download into specified zip file location
-download.file(url = ZenodoLink, destfile = 'data/testclips.zip',method='curl')
-
-# Unzip folder
-exdir <- 'data/'
-utils::unzip(zipfile = 'data/testclips.zip', exdir = exdir )
-
-# Check folder composition
-TestDatapath <- paste(exdir,"testclips",sep='')
-
-# Check folder names
-list.files(TestDatapath)
-
-# Create spectorgram images
-spectrogram_images(
-   trainingBasePath = TestDatapath,
-   outputBasePath = 'data/testimages/',
-   minfreq.khz = 0.4,
-   maxfreq.khz = 1.6,
-   splits = c(0, 0, 1), # Assign proportion to training, validation, or test folders
-   new.sampleratehz = 'NA'
- )
-```
-
-# Train the models
-
-## Training the models using gibbonNetR and evaluating on a test set
-
-``` r
-# Location of spectrogram images for training
-input.data.path <-  'data/trainingimages/'
-
-# Location of spectrogram images for testing
-test.data.path <- 'data/testimages/test/'
-
-# User specified training data label for metadata
-trainingfolder.short <- 'danummulticlassexample'
-
-# We can specify the number of epochs to train here
-epoch.iterations <- c(20)
-
-# Function to train a multi-class CNN
-gibbonNetR::train_CNN_multi(input.data.path=input.data.path,
-                            architecture ='resnet50',
-                            learning_rate = 0.001,
[…]
