+++ pkgdown.yml
@@ -2,8 +2,9 @@ pandoc: 3.1.11
 pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
-  gibbonNetR: gibbonNetR.html
-last_built: 2024-11-24T18:44Z
+  Training multiclass models: Training multiclass models.html
+  UnsupervisedClustering: UnsupervisedClustering.html
+last_built: 2024-11-24T19:18Z
 urls:
   reference: https://denajgibbon.github.io/gibbonNetR/reference
   article: https://denajgibbon.github.io/gibbonNetR/articles
