+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   Training multiclass models: Training multiclass models.html
   UnsupervisedClustering: UnsupervisedClustering.html
-last_built: 2024-11-24T19:27Z
+last_built: 2024-11-24T19:38Z
 urls:
   reference: https://denajgibbon.github.io/gibbonNetR/reference
   article: https://denajgibbon.github.io/gibbonNetR/articles
