+++ vignettes/Automated detection.Rmd
@@ -1,0 +1,57 @@
+---
+title: "3. Automated detection"
+output:
+  rmarkdown::html_vignette:
+    toc: true
+    toc_depth: 1
+    number_sections: true
+    self_contained: yes
+vignette: >
+  %\VignetteIndexEntry{3. Automated detection}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+## We can then deploy the model over longer sound files. 
+```{r, eval=F, warning=FALSE, results='hide'}
+  
+   library(gibbonNetR)
+   
+    # Load data
+   data("TempBinWav")
+
+   # Save in temp directory
+   dir.create(paste(tempdir(),'/MultiDir/Wav/',sep=''),recursive = T, showWarnings = FALSE)
+
+   # Write to temp directory
+   writeWave(TempBinWav,filename = paste(tempdir(),'/MultiDir/Wav/','TempBinWav.wav',sep=''))
+
+   # Find model path
+   # Set model directory
+   trained_models_dir <- system.file("extdata", "trainedresnetmulti/", package = "gibbonNetR")
+
+   # Specify model path
+   ModelPath <- list.files(trained_models_dir,full.names = T)
+
+   # Deploy trained model over sound files
+   deploy_CNN_multi(
+     clip_duration = 12,
+     architecture='resnet18',
+     output_folder = paste(tempdir(),'/MultiDir/Results/Images/',sep=''),
+     output_folder_selections = paste(tempdir(),'/MultiDir/Results/Selections/',sep=''),
+     output_folder_wav = paste(tempdir(),'/MultiDir/Results/Wavs/',sep=''),
+     detect_pattern=NA,
+     top_model_path = ModelPath,
+     path_to_files = paste(tempdir(),'/MultiDir/Wav/',sep=''),
+     downsample_rate = 'NA',
+     save_wav = F,
+     class_names = c('female.gibbon','hornbill.helmeted','hornbill.rhino','long.argus','noise'),
+     noise_category = 'noise',
+     single_class = FALSE,
+     single_class_category = 'female.gibbon',
+     threshold = .25,
+     max_freq_khz = 2
+   )
+   
+```
+
