+++ vignettes/UnsupervisedClustering.Rmd
@@ -13,12 +13,15 @@ vignette: >
 ---
 
 # Extract embeddings
+Embeddings are learned feature representations from CNNs. Instead of doing classification, the CNNs can also be used as feature extractors for spectrogram images. 
+
 Use the pre-trained model to extract embeddings and use unsupervised clustering to identify signals
 
 ```{r, eval=F, warning=FALSE, results='hide'}
 
-ModelPath <- "model_output/_danummulticlassexample_multi_unfrozen_TRUE_/_danummulticlassexample_20_resnet50_model.pt"
-result <- extract_embeddings(test_input="data/examples/test/",
+ModelPath <- "data/model_output/_danummulticlassexample_multi_unfrozen_TRUE_/_danummulticlassexample_1_alexnet_model.pt"
+
+result <- extract_embeddings(test_input="data/testimages/test/",
                                       model_path=ModelPath,
                                      target_class = "female.gibbon")
 ```
@@ -28,6 +31,10 @@ result <- extract_embeddings(test_input="data/examples/test/",
 ```{r, eval=F}
 result$EmbeddingsCombined
 ```
+
+ <img src="unsupervisedclustering.png" align="center" width = "75%" height="75%"/>
+  <figcaption>"Unsupervised clustering with embeddings"<figcaption>
+
 
 ### We can explore the unsupervised clustering results
 
@@ -41,48 +48,5 @@ The confusion matrix results when we use 'hdbscan' to match the target class to 
 
 ```{r, eval=F}
 result$ConfusionMatrix
-```
-
-## We can then deploy the model over longer sound files. 
-```{r, eval=F, warning=FALSE, results='hide'}
-  
-   library(gibbonNetR)
-   
-    # Load data
-   data("TempBinWav")
-
-   # Save in temp directory
-   dir.create(paste(tempdir(),'/MultiDir/Wav/',sep=''),recursive = T, showWarnings = FALSE)
-
-   # Write to temp directory
-   writeWave(TempBinWav,filename = paste(tempdir(),'/MultiDir/Wav/','TempBinWav.wav',sep=''))
-
-   # Find model path
-   # Set model directory
-   trained_models_dir <- system.file("extdata", "trainedresnetmulti/", package = "gibbonNetR")
-
-   # Specify model path
-   ModelPath <- list.files(trained_models_dir,full.names = T)
-
-   # Deploy trained model over sound files
-   deploy_CNN_multi(
-     clip_duration = 12,
-     architecture='resnet18',
-     output_folder = paste(tempdir(),'/MultiDir/Results/Images/',sep=''),
-     output_folder_selections = paste(tempdir(),'/MultiDir/Results/Selections/',sep=''),
-     output_folder_wav = paste(tempdir(),'/MultiDir/Results/Wavs/',sep=''),
-     detect_pattern=NA,
-     top_model_path = ModelPath,
-     path_to_files = paste(tempdir(),'/MultiDir/Wav/',sep=''),
-     downsample_rate = 'NA',
-     save_wav = F,
-     class_names = c('female.gibbon','hornbill.helmeted','hornbill.rhino','long.argus','noise'),
-     noise_category = 'noise',
-     single_class = FALSE,
-     single_class_category = 'female.gibbon',
-     threshold = .25,
-     max_freq_khz = 2
-   )
-   
 ```
 
