+++ pkgdown.yml
@@ -2,9 +2,11 @@ pandoc: 3.1.11
 pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
+  Automated detection: Automated detection.html
+  Training binary models: Training binary models.html
   Training multiclass models: Training multiclass models.html
   UnsupervisedClustering: UnsupervisedClustering.html
-last_built: 2024-11-24T19:38Z
+last_built: 2024-11-24T19:59Z
 urls:
   reference: https://denajgibbon.github.io/gibbonNetR/reference
   article: https://denajgibbon.github.io/gibbonNetR/articles
