+++ pkgdown.yml
@@ -2,11 +2,11 @@ pandoc: 3.1.11
 pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
-  Automated detection: Automated detection.html
-  Training binary models: Training binary models.html
-  Training multiclass models: Training multiclass models.html
-  UnsupervisedClustering: UnsupervisedClustering.html
-last_built: 2024-11-24T19:59Z
+  01-Training multiclass models: 01-Training multiclass models.html
+  02-Training binary models: 02-Training binary models.html
+  03-Automated detection: 03-Automated detection.html
+  04-UnsupervisedClustering: 04-UnsupervisedClustering.html
+last_built: 2024-11-24T20:05Z
 urls:
   reference: https://denajgibbon.github.io/gibbonNetR/reference
   article: https://denajgibbon.github.io/gibbonNetR/articles
