### 2024-11-24 c911724
Update vignettes.
