+++ vignettes/a02-Training binary models.Rmd
@@ -28,7 +28,6 @@ knitr::opts_chunk$set(
 options(width = 100, max.print = 100)
 ```
 
-# NOTE THIS IS THE SAME AS MULTI
 # Data Preparation
 
 ## First load the package
@@ -41,17 +40,17 @@ To follow best practices we want to use training and test data from different pa
 
 ```{r eval=FALSE}
 # Link to training clips on Zenodo
-ZenodoLink <- 'https://zenodo.org/records/14212086/files/trainingclips.zip?download=1'
+ZenodoLink <- 'https://zenodo.org/records/14213067/files/trainingclipsbinary.zip?download=1'
 
 # Download into specified zip file location
-download.file(url = ZenodoLink, destfile = 'data/trainingclips.zip',method='curl')
+download.file(url = ZenodoLink, destfile = 'data/trainingclipsbinary.zip',method='curl')
 
 # Unzip folder
 exdir <- 'data/'
-utils::unzip(zipfile = 'data/trainingclips.zip', exdir = exdir )
+utils::unzip(zipfile = 'data/trainingclipsbinary.zip', exdir = exdir )
 
 # Check folder composition
-TrainingDatapath <- paste(exdir,"trainingclips",sep='')
+TrainingDatapath <- paste(exdir,"trainingclipsbinary",sep='')
 
 # Check folder names
 list.files(TrainingDatapath)
@@ -59,7 +58,7 @@ list.files(TrainingDatapath)
 # Create spectrogram images
 spectrogram_images(
    trainingBasePath = TrainingDatapath,
-   outputBasePath = 'data/trainingimages/',
+   outputBasePath = 'data/trainingimagesbinary/',
    minfreq.khz = 0.4,
    maxfreq.khz = 1.6,
    random=FALSE,
@@ -70,29 +69,23 @@ spectrogram_images(
 
 ```
 
-```{r, echo=FALSE, out.width="50%", fig.cap="A sample of the spectrogram images."}
-#knitr::include_graphics("spectro.png")
-```
-
-<img src="spectro.png" align="center" width = "50%" height="50%"/>
-
 ## Download example test files on Zenodo and convert to spectrogram images
 The "spectrogram_images" function below puts 100% of the test images into the test folder. 
 ```{r eval = FALSE}
 library(gibbonNetR)
 
 # Link to training clips on Zenodo
-ZenodoLink <- 'https://zenodo.org/records/14212086/files/testclips.zip?download=1'
+ZenodoLink <- 'https://zenodo.org/records/14213067/files/testclipsbinary.zip?download=1'
 
 # Download into specified zip file location
-download.file(url = ZenodoLink, destfile = 'data/testclips.zip',method='curl')
+download.file(url = ZenodoLink, destfile = 'data/testclipsbinary.zip',method='curl')
 
 # Unzip folder
 exdir <- 'data/'
-utils::unzip(zipfile = 'data/testclips.zip', exdir = exdir )
+utils::unzip(zipfile = 'data/testclipsbinary.zip', exdir = exdir )
 
 # Check folder composition
-TestDatapath <- paste(exdir,"testclips",sep='')
+TestDatapath <- paste(exdir,"testclipsbinary",sep='')
 
 # Check folder names
 list.files(TestDatapath)
@@ -100,7 +93,7 @@ list.files(TestDatapath)
 # Create spectorgram images
 spectrogram_images(
    trainingBasePath = TestDatapath,
-   outputBasePath = 'data/testimages/',
+   outputBasePath = 'data/testimagesbinary/',
    minfreq.khz = 0.4,
    maxfreq.khz = 1.6,
    splits = c(0, 0, 1), # Assign proportion to training, validation, or test folders
@@ -110,37 +103,44 @@ spectrogram_images(
 ```
 
 # Model training
[…]
