+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   a02-Training binary models: a02-Training binary models.html
   a03-Automated detection: a03-Automated detection.html
   a04-UnsupervisedClustering: a04-UnsupervisedClustering.html
-last_built: 2024-11-24T22:06Z
+last_built: 2024-11-24T22:09Z
 urls:
   reference: https://denajgibbon.github.io/gibbonNetR/reference
   article: https://denajgibbon.github.io/gibbonNetR/articles
