### 2024-12-18 4f38e46
Update to using 'pROC' for AUC calculations for training both binary and multi.
