+++ R/train_CNN_multi.R
@@ -422,7 +422,7 @@ train_CNN_multi <- function(input.data.path, test.data, architecture,
         ifelse(outputTableMultiSub$ActualClass==UniqueClasses[b],UniqueClasses[b],noise.category)
 
       binarylabels <- ifelse(outputTableMulti$ActualClass==UniqueClasses[b],1,0)
-      roc_result <- pROC::roc(response=binarylabels, predictor = as.numeric(Probability[, UniqueClasses[b]]),direction = "<")
+      #roc_result <- pROC::roc(response=binarylabels, predictor = as.numeric(Probability[, UniqueClasses[b]]),direction = "<")
 
       for (threshold in thresholds) {
         MultiPredictedClass <- ifelse((outputTableMultiSub$Probability > threshold ), UniqueClasses[b], noise.category)
@@ -453,13 +453,14 @@ train_CNN_multi <- function(input.data.path, test.data, architecture,
           "CNN Architecture"
         )
 
-          # ROCRpred <- ROCR::prediction(predictions = outputTableMultiSub$Probability, labels = outputTableMultiSub$ActualClass)
-        # AUCval <- ROCR::performance(ROCRpred, 'auc')
-        TempRowMulti$AUC <- as.numeric(roc_result$auc)
+        ROCRpred <- ROCR::prediction(predictions = outputTableMultiSub$Probability, labels = outputTableMultiSub$ActualClass)
+        AUCval <- ROCR::performance(ROCRpred, 'auc')
+        TempRowMulti$AUC <- as.numeric(AUCval@y.values)
         TempRowMulti$Threshold <- as.character(threshold)
         TempRowMulti$Frozen <- unfreeze.param
         TempRowMulti$Class <- UniqueClasses[b]
         TempRowMulti$Class <- as.factor(TempRowMulti$Class)
+        TempRowMulti$TestDataPath <-  test.data
         CombinedTempRow <- rbind.data.frame(CombinedTempRow, TempRowMulti)
       }
     }
