+++ man/evaluate_trainedmodel_performance.Rd
@@ -32,5 +32,4 @@ Given trained models and a set of images, this function evaluates the performanc
 \examples{
 {
 }
-
 }
