+++ vignettes/a02-Training multiclass models.Rmd
@@ -7,7 +7,7 @@ output:
     number_sections: true
     self_contained: yes
 vignette: >
-  %\VignetteIndexEntry{1. Multiclass models}
+  %\VignetteIndexEntry{2. Multiclass models}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
