+++ vignettes/a01-OverviewAndGettingStarted.Rmd
@@ -7,7 +7,7 @@ output:
     number_sections: true
     self_contained: yes
 vignette: >
-  %\VignetteIndexEntry{1. Multiclass models}
+  %\VignetteIndexEntry{1. Overview and Getting Started}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
@@ -27,3 +27,6 @@ knitr::opts_chunk$set(
 
 options(width = 100, max.print = 100)
 ```
+
+
+
