+++ vignettes/a03-Training binary models.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "2. Binary models"
+title: "3. Binary models"
 output:
   rmarkdown::html_vignette:
     toc: true
@@ -7,7 +7,7 @@ output:
     number_sections: true
     self_contained: yes
 vignette: >
-  %\VignetteIndexEntry{2. Binary models}
+  %\VignetteIndexEntry{3. Binary models}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
