### 2025-02-08 744bf82
Fix tests and add package-level documentation.
