+++ man/gibbonNetR-package.Rd
@@ -1,0 +1,22 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/gibbonNetR-package.R
+\docType{package}
+\name{gibbonNetR-package}
+\alias{gibbonNetR}
+\alias{gibbonNetR-package}
+\title{gibbonNetR: gibbonNetR: R Package for the Use of CNNs and Transfer Learning on Acoustic Data}
+\description{
+R Package for the Use of CNNs and Transfer Learning on Acoustic Data.
+}
+\seealso{
+Useful links:
+\itemize{
+  \item \url{https://denajgibbon.github.io/gibbonNetR/}
+}
+
+}
+\author{
+\strong{Maintainer}: Dena J. Clink \email{dena.clink@cornell.edu} (\href{https://orcid.org/0000-0003-0363-5581}{ORCID})
+
+}
+\keyword{internal}
