+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   a03-Training binary models: a03-Training binary models.html
   a04-Automated detection: a04-Automated detection.html
   a05-UnsupervisedClustering: a05-UnsupervisedClustering.html
-last_built: 2025-02-08T16:28Z
+last_built: 2025-02-08T16:32Z
 urls:
   reference: https://denajgibbon.github.io/gibbonNetR/reference
   article: https://denajgibbon.github.io/gibbonNetR/articles
