+++ R/evaluate_trainedmodel_performance_multi.R
@@ -12,42 +12,25 @@
 #' @return Invisible NULL. The performance scores are written to the specified output directory.
 #' @importFrom stringr str_split_fixed str_detect
 #' @examples {
-# if (.Platform$OS.type == "windows") {
-#   # Set directory paths for trained models and test images
-#
-#   trained_models_dir <- system.file("extdata", "trainedresnetmulti\\", package = "gibbonNetR")
-#   image_data_dir <- system.file("extdata", "multiclass\\test\\", package = "gibbonNetR")
-#   class_names <-  c('female.gibbon','hornbill.helmeted','hornbill.rhino','long.argus','noise')
-#   # Evaluate the performance of the trained models using the test images
-#   evaluate_trainedmodel_performance_multi(trained_models_dir=trained_models_dir,
-#                                           class_names=class_names,
-#                                           image_data_dir=image_data_dir,
-#                                           output_dir= paste(tempdir(), '\\data\\',sep=''),
-#                                           noise.category = "noise")
-#   # Find the location of saved evaluation files
-#   CSVName <- list.files(paste(tempdir(), '\\data\\',sep=''), recursive = TRUE, full.names = TRUE)
-#
-#   # Check the output of the first file
-#   head(read.csv(CSVName[1]))
-#
-# } else {
-#
-#   trained_models_dir <- system.file("extdata", "trainedresnetmulti/", package = "gibbonNetR")
-#   image_data_dir <- system.file("extdata", "multiclass/test/", package = "gibbonNetR")
-#   class_names <-  c('female.gibbon','hornbill.helmeted','hornbill.rhino','long.argus','noise')
-#   # Evaluate the performance of the trained models using the test images
-#   evaluate_trainedmodel_performance_multi(trained_models_dir=trained_models_dir,
-#                                           class_names=class_names,
-#                                           image_data_dir=image_data_dir,
-#                                           output_dir= paste(tempdir(), '/data/'),
-#                                           noise.category = "noise")
-#   # Find the location of saved evaluation files
-#   CSVName <- list.files(paste(tempdir(), '/data/'), recursive = TRUE, full.names = TRUE)
-#
-#   # Check the output of the first file
-#   head(read.csv(CSVName[1]))
-#
-# }
+#' # Set directory paths for trained models and test images
+#' trained_models_dir <- system.file("extdata", "trainedresnetmulti", package = "gibbonNetR")
+#' image_data_dir <- system.file("extdata", "multiclass", "test", package = "gibbonNetR")
+#' class_names <- c('female.gibbon', 'hornbill.helmeted', 'hornbill.rhino', 'long.argus', 'noise')
+#'
+#' # Evaluate the performance of the trained models using the test images
+#' evaluate_trainedmodel_performance_multi(
+#'   trained_models_dir = trained_models_dir,
+#'   class_names = class_names,
+#'   image_data_dir = image_data_dir,
+#'   output_dir = file.path(tempdir(), 'data/'),
+#'   noise.category = "noise"
+#' )
+#'
+#' # Find the location of saved evaluation files
+#' CSVName <- list.files(file.path(tempdir(), 'data'), recursive = TRUE, full.names = TRUE)
+#'
+#' # Check the output of the first file
+#' head(read.csv(CSVName[1]))
 #'}
 #'
 #' @importFrom purrr %>%
