+++ R/evaluate_trainedmodel_performance.R
@@ -12,46 +12,25 @@
 #' @importFrom stringr str_split_fixed str_detect
 #' @importFrom purrr %>%
 #' @examples {
-# if (.Platform$OS.type == "windows") {
-#   # Set directory paths for trained models and test images
-#
-#   trained_models_dir <- system.file("extdata", "trainedresnetbinary\\", package = "gibbonNetR")
-#   image_data_dir <- system.file("extdata", "binary\\test\\", package = "gibbonNetR")
-#
-#   # Evaluate the performance of the trained models using the test images
-#   evaluate_trainedmodel_performance(trained_models_dir = trained_models_dir,
-#                                     image_data_dir = image_data_dir,
-#                                     output_dir = paste(tempdir(), '\\data\\',sep=''),
-#                                     positive.class = 'Gibbons',  #' Label for positive class
-#                                     negative.class = 'Noise')    #' Label for negative class
-#
-#   # Find the location of saved evaluation files
-#   CSVName <- list.files(paste(tempdir(), '\\data\\',sep=''), recursive = TRUE, full.names = TRUE)
-#
-#   # Check the output of the first file
-#   head(read.csv(CSVName[1]))
-#
-# } else  {
-#
-#   # Set directory paths for trained models and test images
-#
-#   trained_models_dir <- system.file("extdata", "trainedresnetbinary/", package = "gibbonNetR")
-#   image_data_dir <- system.file("extdata", "binary/test/", package = "gibbonNetR")
-#
-#   # Evaluate the performance of the trained models using the test images
-#   evaluate_trainedmodel_performance(trained_models_dir = trained_models_dir,
-#                                     image_data_dir = image_data_dir,
-#                                     output_dir = paste(tempdir(), '/data/'),
-#                                     positive.class = 'Gibbons',  #' Label for positive class
-#                                     negative.class = 'Noise')    #' Label for negative class
-#
-#   # Find the location of saved evaluation files
-#   CSVName <- list.files(paste(tempdir(), '/data/'), recursive = TRUE, full.names = TRUE)
-#
-#   # Check the output of the first file
-#   head(read.csv(CSVName[1]))
-#
-# }
+#' # Set directory paths for trained models and test images
+#' trained_models_dir <- system.file("extdata", "trainedresnetbinary", package = "gibbonNetR")
+#' image_data_dir <- system.file("extdata", "binary", "test", package = "gibbonNetR")
+#'
+#' # Evaluate the performance of the trained models using the test images
+#' evaluate_trainedmodel_performance(
+#'   trained_models_dir = trained_models_dir,
+#'   image_data_dir = image_data_dir,
+#'   output_dir = file.path(tempdir(), 'data/'),
+#'   positive.class = 'Gibbons',  # Label for positive class
+#'   negative.class = 'Noise'    # Label for negative class
+#' )
+#'
+#' # Find the location of saved evaluation files
+#' CSVName <- list.files(file.path(tempdir(), 'data'), recursive = TRUE, full.names = TRUE)
+#'
+#' # Check the output of the first file
+#' head(read.csv(CSVName[1]))
+#'
 #'}
 
 #' @export

[…]
