### 2025-02-08 DenaJGibbon/gibbonNetR#5 @DenaJGibbon
Package-level documentation added and replaced paste with file.path() on 02/08/25.
