### 2025-02-08 DenaJGibbon/gibbonNetR#3 @DenaJGibbon
License text manually moved.
