### 2025-02-08 bacb484
Push function example documentation.
