+++ man/deploy_CNN_multi.Rd
@@ -71,75 +71,39 @@ This function processes sound data from a directory, extracts sound clips, conve
 }
 \examples{
 {
-if (.Platform$OS.type == "windows") {
+# Load data
+data("TempBinWav")
 
-  # Load data
-  data("TempBinWav")
+# Create necessary directories
+dir.create(file.path(tempdir(), "MultiDir", "Wav"), recursive = TRUE, showWarnings = FALSE)
 
-  dir.create(paste(tempdir(),'\\\\MultiDir\\\\Wav\\\\',sep=''),recursive = TRUE, showWarnings = FALSE)
+# Write to temp directory
+writeWave(TempBinWav, filename = file.path(tempdir(), "MultiDir", "Wav", "TempBinWav.wav"))
 
-  # Write to temp directory
-  writeWave(TempBinWav,filename = paste(tempdir(),'\\\\MultiDir\\\\Wav\\\\','TempBinWav.wav',sep=''))
+# Set model directory
+trained_models_dir <- system.file("extdata", "trainedresnetmulti", package = "gibbonNetR")
 
-  #' Set model directory
-  trained_models_dir <- system.file("extdata", "trainedresnetmulti\\\\", package = "gibbonNetR")
+# Specify model path
+ModelPath <- list.files(trained_models_dir, full.names = TRUE)
 
-  #' Specify model path
-  ModelPath <- list.files(trained_models_dir,full.names = TRUE)
+# Deploy trained model over sound files
+deploy_CNN_multi(
+  clip_duration = 12,
+  architecture = "resnet18",
+  output_folder = file.path(tempdir(), "MultiDir", "Results", "Images"),
+  output_folder_selections = file.path(tempdir(), "MultiDir", "Results", "Selections"),
+  output_folder_wav = file.path(tempdir(), "MultiDir", "Results", "Wavs"),
+  detect_pattern = NA,
+  top_model_path = ModelPath,
+  path_to_files = file.path(tempdir(), "MultiDir", "Wav"),
+  downsample_rate = "NA",
+  save_wav = FALSE,
+  class_names = c("female.gibbon", "hornbill.helmeted", "hornbill.rhino", "long.argus", "noise"),
+  noise_category = "noise",
+  single_class = FALSE,
+  threshold = 0.25,
+  max_freq_khz = 2
+)
 
-  #' Deploy trained model over sound files
-  deploy_CNN_multi(
-    clip_duration = 12,
-    architecture='resnet18',
-    output_folder = paste(tempdir(),'\\\\MultiDir\\\\Results\\\\Images\\\\',sep=''),
-    output_folder_selections = paste(tempdir(),'\\\\MultiDir\\\\Results\\\\Selections\\\\',sep=''),
-    output_folder_wav = paste(tempdir(),'\\\\MultiDir\\\\Results\\\\Wavs\\\\',sep=''),
-    detect_pattern=NA,
-    top_model_path = ModelPath,
-    path_to_files = paste(tempdir(),'\\\\MultiDir\\\\Wav\\\\',sep=''),
-    downsample_rate = 'NA',
-    save_wav = F,
-    class_names = c('female.gibbon','hornbill.helmeted','hornbill.rhino','long.argus','noise'),
-    noise_category = 'noise',
-    single_class = FALSE,
-    threshold = .25,
-    max_freq_khz = 2
-  )
-} else  {
-
-  # Load data
-  data("TempBinWav")
-
-  dir.create(paste(tempdir(),'/MultiDir/Wav/',sep=''),recursive = TRUE, showWarnings = FALSE)
-
-  # Write to temp directory
[…]
