+++ man/deploy_CNN_binary.Rd
@@ -61,73 +61,37 @@ This function processes sound data from a directory, extracts sound clips, conve
 }
 \examples{
 {
-if (.Platform$OS.type == "windows") {
-
   #' Load data
   data("TempBinWav")
 
-  dir.create(paste(tempdir(),'\\\\BinaryDir\\\\Wav\\\\',sep=''),recursive = TRUE, showWarnings = T)
+  #' Create necessary directories
+  dir.create(file.path(tempdir(), "BinaryDir", "Wav"), recursive = TRUE, showWarnings = TRUE)
 
   #' Write to temp directory
-  writeWave(TempBinWav,filename = paste(tempdir(),'\\\\BinaryDir\\\\Wav\\\\','TempBinWav.wav',sep=''))
+  writeWave(TempBinWav, filename = file.path(tempdir(), "BinaryDir", "Wav", "TempBinWav.wav"))
 
   #' Set model directory
-  trained_models_dir <- system.file("extdata", "trainedresnetbinary\\\\", package = "gibbonNetR")
+  trained_models_dir <- system.file("extdata", "trainedresnetbinary", package = "gibbonNetR")
 
   #' Specify model path
-  ModelPath <- list.files(trained_models_dir,full.names =TRUE)
+  ModelPath <- list.files(trained_models_dir, full.names = TRUE)
 
-
-  deploy_CNN_binary (
+  #' Run model
+  deploy_CNN_binary(
     clip_duration = 12,
-    architecture='alexnet',
-    output_folder = paste(tempdir(),'\\\\BinaryDir\\\\Results\\\\Images\\\\',sep=''),
-    output_folder_selections = paste(tempdir(),'\\\\BinaryDir\\\\Results\\\\Selections\\\\',sep=''),
-    output_folder_wav = paste(tempdir(),'\\\\BinaryDir\\\\Results\\\\Wavs\\\\',sep=''),
-    detect_pattern=NA,
+    architecture = "alexnet",
+    output_folder = file.path(tempdir(), "BinaryDir", "Results", "Images"),
+    output_folder_selections = file.path(tempdir(), "BinaryDir", "Results", "Selections"),
+    output_folder_wav = file.path(tempdir(), "BinaryDir", "Results", "Wavs"),
+    detect_pattern = NA,
     top_model_path = ModelPath,
-    path_to_files = paste(tempdir(),'\\\\BinaryDir\\\\Wav\\\\',sep=''),
-    downsample_rate = 'NA',
+    path_to_files = file.path(tempdir(), "BinaryDir", "Wav"),
+    downsample_rate = "NA",
     threshold = 0.5,
-    save_wav = F,
-    positive.class = 'Gibbons',
-    negative.class = 'Noise',
+    save_wav = FALSE,
+    positive.class = "Gibbons",
+    negative.class = "Noise",
     max_freq_khz = 2
   )
-
-
-} else  {
-  #' Load data
-  data("TempBinWav")
-
-  dir.create(paste(tempdir(),'/BinaryDir/Wav/',sep=''),recursive = TRUE, showWarnings = FALSE)
-
-  #' Write to temp directory
-  writeWave(TempBinWav,filename = paste(tempdir(),'/BinaryDir/Wav/','TempBinWav.wav',sep=''))
-
-  #' Set model directory
-  trained_models_dir <- system.file("extdata", "trainedresnetbinary/", package = "gibbonNetR")
-
-  #' Specify model path
-  ModelPath <- list.files(trained_models_dir,full.names =TRUE)
-
-
-  deploy_CNN_binary (
-    clip_duration = 12,
-    architecture='alexnet',
[…]
