### 2025-02-08 DenaJGibbon/gibbonNetR#4 @DenaJGibbon
I am going to mark this as closed, as the site has been created. I am still working on the other suggestions regarding explaining the workflow, and that will be done on the pkgdown site. This was such a great suggestion! Thank you.
