### 2025-02-08 200f6d8
Update spectrogram example.
