+++ R/spectrogram_images.R
@@ -10,106 +10,107 @@
 #' @return Invisible NULL
 #'
 #' @examples
+#' @examples
 #' {
-#' #' # Load data
+#' #' # Load the 'TempBinWav' data
 #' data("TempBinWav")
 #'
-#' # Define the output directory
+#' #' # Define the output directory for saving .wav files
 #' output.dir <- file.path(tempdir(), 'MultiDir', 'Soundfiles', 'Noise')
 #'
-#' # Create the output directory
+#' #' # Create the output directory if it doesn't already exist
 #' dir.create(output.dir, recursive = TRUE, showWarnings = FALSE)
 #'
-#' # Define the intervals for cutting the wave
+#' #' # Define the intervals for cutting the wave (from 1 to 30 with a step of 5)
 #' cutwave.list <- seq(1, 30, 5)
 #'
-#' # Extract subsamples from the waveform
+#' #' # Extract subsamples from the waveform by looping through the cutwave intervals
 #' subsamps <- lapply(1:(length(cutwave.list) - 1),
 #'                    function(i)
 #'                      extractWave(
-#'                        TempBinWav,
-#'                        from = cutwave.list[i],
-#'                        to = cutwave.list[i + 1],
-#'                        xunit = "time",
-#'                        plot = FALSE,
-#'                        output = "Wave"
+#'                        TempBinWav,         # Input waveform data
+#'                        from = cutwave.list[i],  # Start point of the cut interval
+#'                        to = cutwave.list[i + 1], # End point of the cut interval
+#'                        xunit = "time",     # Specify the unit as time
+#'                        plot = FALSE,       # Don't plot the waveform
+#'                        output = "Wave"     # Specify the output type (Wave)
 #'                      ))
 #'
-#' # Write the extracted subsamples to .wav files
+#' #' # Write the extracted subsamples to .wav files in the defined output directory
 #' lapply(1:length(subsamps),
 #'        function(i)
 #'          writeWave(
-#'            subsamps[[i]],
-#'            filename = file.path(output.dir, paste('temp_', i, '_', '.wav', sep = '')),
-#'            extensible = FALSE
+#'            subsamps[[i]],  # Subsample to save
+#'            filename = file.path(output.dir, paste('temp_', i, '_', '.wav', sep = '')),  # File path with a dynamic naming pattern
+#'            extensible = FALSE  # Disable extensible wave format
 #'          )
 #' )
 #'
-#' # List the files in the output directory
+#' #' # List all the files in the output directory
 #' list.files(output.dir)
 #'
-#' # Generate spectrogram images
+#' #' # Generate spectrogram images for sound files located in 'Soundfiles' directory
 #' spectrogram_images(
-#'   trainingBasePath =  file.path(tempdir(), 'MultiDir/Soundfiles/'),
-#'   outputBasePath = file.path(tempdir(), 'MultiDir/Soundfiles/', 'Spectro'),
-#'   splits = c(0.5, 0.5, 0.0),
-#'   new.sampleratehz = 'NA'
+#'   trainingBasePath = file.path(tempdir(), 'MultiDir/Soundfiles/'),  # Base path for input sound files
+#'   outputBasePath = file.path(tempdir(), 'MultiDir/', 'Spectro'),  # Base path for saving generated spectrogram images
+#'   splits = c(0.5, 0.5, 0.0),  # Split the data into training (50%), validation (50%), and no test data (0%)
+#'   new.sampleratehz = 'NA'  # No change to the sampling rate
 #' )
 #'
-#' # List the images generated
-#' ListImages <- list.files(file.path(tempdir(), 'MultiDir', 'Spectro'), recursive = TRUE)
-#'
[…]
