+++ .github/workflows/pkgdown.yaml
@@ -33,7 +33,7 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: any::torch, any::pkgdown, local::.
+          extra-packages: any::pkgdown, local::.
           needs: website
 
       - name: Build site
