+++ vignettes/a01-OverviewAndGettingStarted.Rmd
@@ -28,5 +28,53 @@ knitr::opts_chunk$set(
 options(width = 100, max.print = 100)
 ```
 
+```{r, eval = TRUE, echo = FALSE, message=FALSE}
+# Install required package if not installed
+library(DiagrammeR)
+
+flowchart <- "
+digraph G {
+  graph [layout = dot, rankdir = TB]
+
+  # Nodes: Scripts as rectangles, key decision points as diamonds
+  spectrogram [label = 'Generate Spectrograms\\n(spectrogram_images.R)', shape = rectangle, style = filled, fillcolor = lightblue]
+    prepare_data [label = 'Prepare Data', shape = diamond, style = filled, fillcolor = white]
+    
+  train_model [label = 'Train Model', shape = diamond, style = filled, fillcolor = white]
+  train_binary [label = 'Train Binary CNN\\n(train_CNN_binary.R)', shape = rectangle, style = filled, fillcolor = green]
+  train_multi [label = 'Train Multi CNN\\n(train_CNN_multi.R)', shape = rectangle, style = filled, fillcolor = green]
+
+  eval_model [label = 'Evaluate Performance', shape = diamond, style = filled, fillcolor = white]
+  eval_binary [label = 'Evaluate Binary CNN\\n(evaluate_trainedmodel_performance.R)', shape = rectangle, style = filled, fillcolor = pink]
+  eval_multi [label = 'Evaluate Multi CNN\\n(evaluate_trainedmodel_performance_multi.R)', shape = rectangle, style = filled, fillcolor = pink]
+  
+  best_performance [label = 'Get Best Performance\\n(get_best_performance.R)', shape = rectangle, style = filled, fillcolor = pink]
+
+  deploy_model [label = 'Deploy Model', shape = diamond, style = filled, fillcolor = white]
+  deploy_binary [label = 'Deploy Binary CNN\\n(deploy_CNN_binary.R)', shape = rectangle, style = filled, fillcolor = orange]
+  deploy_multi [label = 'Deploy Multi CNN\\n(deploy_CNN_multi.R)', shape = rectangle, style = filled, fillcolor = orange]
+
+data_viz [label = 'Extract embeddings + \n unsupervised clustering', shape = diamond, style = filled, fillcolor = white]
+  extract_embed [label = 'Extract Embeddings (Optional Step)\\n(extract_embeddings.R)', shape = rectangle, style = filled, fillcolor = cyan]
+
+  # Define Dependencies with Decision Points
+  prepare_data -> spectrogram
+  spectrogram -> train_model
+  train_model -> { train_binary train_multi }
+
+  { train_binary train_multi } -> eval_model
+  eval_model -> { eval_binary eval_multi best_performance }
+
+  { eval_binary eval_multi best_performance } -> deploy_model
+  deploy_model -> { deploy_binary deploy_multi }
+
+  { deploy_binary deploy_multi } -> data_viz
+  
+  data_viz -> extract_embed
+}
+"
+
+DiagrammeR::grViz(flowchart)
 
 
+```
