+++ vignettes/a01-OverviewAndGettingStarted.Rmd
@@ -14,20 +14,21 @@ vignette: >
 
 ```{r, eval = TRUE, echo = FALSE, message=FALSE}
 # global option chunks
-knitr::opts_chunk$set(
-  fig.width = 5, 
-  fig.height = 3.5,
-  dpi = 70,
-  comment = "",
-  out.width = "80%",
-  fig.align = "center",
-  message = TRUE,
-  warning = TRUE
-)
-
-options(width = 100, max.print = 100)
+# knitr::opts_chunk$set(
+#   fig.width = 5, 
+#   fig.height = 3.5,
+#   dpi = 70,
+#   comment = "",
+#   out.width = "80%",
+#   fig.align = "center",
+#   message = TRUE,
+#   warning = TRUE
+# )
+# 
+# options(width = 100, max.print = 100)
 ```
 
+# A flowchart overview of the 'gibbonNetR' workflow
 ```{r, eval = TRUE, echo = FALSE, message=FALSE}
 if (!requireNamespace("DiagrammeR", quietly = TRUE)) {
   install.packages("DiagrammeR")
