+++ .github/workflows/pkgdown.yaml
@@ -36,6 +36,12 @@ jobs:
           extra-packages: any::pkgdown, local::.
           needs: website
 
+      # Install Lantern explicitly
+      - name: Install Lantern dependency for torch
+        run: |
+          R -e "install.packages('torch')"
+          R -e "torch::install_torch()"
+
       - name: Build site
         run: pkgdown::build_site_github_pages(new_process = FALSE, install = FALSE)
         shell: Rscript {0}
