### 2025-02-09 5adb0a7
Update spectro function.
