+++ R/spectrogram_images.R
@@ -102,11 +102,12 @@
 #' }
 #'
 #' #' # Display the images after normalizing them, with titles for each class
-#' images %>%
+#' images <- images %>%
 #'   purrr::array_tree(1) %>%
 #'   purrr::set_names(class_names) %>%
 #'   purrr::map(~ as.raster(normalize_pixel_values(.x))) %>%
 #'   purrr::iwalk(~ {plot(.x); title(.y)})  # Plot the image with its respective class name as title
+#'   print(images)
 #' }
 #'
 #' @importFrom tuneR readWave
