+++ man/spectrogram_images.Rd
@@ -129,11 +129,12 @@ normalize_pixel_values <- function(image) {
 }
 
 #' # Display the images after normalizing them, with titles for each class
-images \%>\%
+images <- images \%>\%
   purrr::array_tree(1) \%>\%
   purrr::set_names(class_names) \%>\%
   purrr::map(~ as.raster(normalize_pixel_values(.x))) \%>\%
   purrr::iwalk(~ {plot(.x); title(.y)})  # Plot the image with its respective class name as title
+  print(images)
 }
 
 }
