### 2025-02-09 3bf5caf
Revert spectro function
