+++ pkgdown.yml
@@ -1,13 +1,15 @@
 pandoc: 3.1.11
-pkgdown: 2.1.1
+pkgdown: 2.0.9
 pkgdown_sha: ~
 articles:
-  a01-OverviewAndGettingStarted: a01-OverviewAndGettingStarted.html
+  a01a-Overview: a01a-Overview.html
+  a01b-GettingStarted: a01b-GettingStarted.html
   a02-Training multiclass models: a02-Training multiclass models.html
   a03-Training binary models: a03-Training binary models.html
   a04-Automated detection: a04-Automated detection.html
   a05-UnsupervisedClustering: a05-UnsupervisedClustering.html
-last_built: 2025-02-10T12:53Z
+last_built: 2025-02-10T13:25Z
 urls:
   reference: https://denajgibbon.github.io/gibbonNetR/reference
   article: https://denajgibbon.github.io/gibbonNetR/articles
+
