+++ README.Rmd
@@ -17,7 +17,7 @@ knitr::opts_knit$set(root.dir = '/Users/denaclink/Desktop/RStudioProjects/gibbon
 
 # Overview
 
-This README provides code for training and testing the performance of different convolutional neural network model architectures on spectrogram images.
+This README provides code for training, testing, and deployings, different convolutional neural network model architectures on acoustic data using spectrogram images.
 
 # Usage
 
