+++ vignettes/a03-Training binary models.Rmd
@@ -28,81 +28,7 @@ knitr::opts_chunk$set(
 options(width = 100, max.print = 100)
 ```
 
-# Data Preparation
-
-## First load the package
-```{r message=FALSE}
-library(gibbonNetR)
-```
-
-## We create spectrogram images of the sound files. 
-To follow best practices we want to use training and test data from different passive acoustic monitoring arrays These clips are from Danum Valley Conservation Area, Sabah, Malaysia, but are taken from two different arrays. The "spectrogram_images" function takes folders organized and labeled by signal type, and creates training, validation, or test folders using the "splits" specified in the function call. In the example below 70% of the data is used for training and 30% for validation.   
-
-```{r eval=FALSE}
-# Link to training clips on Zenodo
-ZenodoLink <- 'https://zenodo.org/records/14213067/files/trainingclipsbinary.zip?download=1'
-
-# Download into specified zip file location
-download.file(url = ZenodoLink, destfile = 'data/trainingclipsbinary.zip',method='curl')
-
-# Unzip folder
-exdir <- 'data/'
-utils::unzip(zipfile = 'data/trainingclipsbinary.zip', exdir = exdir )
-
-# Check folder composition
-TrainingDatapath <- paste(exdir,"trainingclipsbinary",sep='')
-
-# Check folder names
-list.files(TrainingDatapath)
-
-# Create spectrogram images
-spectrogram_images(
-   trainingBasePath = TrainingDatapath,
-   outputBasePath = 'data/trainingimagesbinary/',
-   minfreq.khz = 0.4,
-   maxfreq.khz = 1.6,
-   random=FALSE,
-   splits = c(0.7, 0.3, 0), # Assign proportion to training, validation, or test folders
-   new.sampleratehz = 'NA'
- )
-
-
-```
-
-## Download example test files on Zenodo and convert to spectrogram images
-The "spectrogram_images" function below puts 100% of the test images into the test folder. 
-```{r eval = FALSE}
-library(gibbonNetR)
-
-# Link to training clips on Zenodo
-ZenodoLink <- 'https://zenodo.org/records/14213067/files/testclipsbinary.zip?download=1'
-
-# Download into specified zip file location
-download.file(url = ZenodoLink, destfile = 'data/testclipsbinary.zip',method='curl')
-
-# Unzip folder
-exdir <- 'data/'
-utils::unzip(zipfile = 'data/testclipsbinary.zip', exdir = exdir )
-
-# Check folder composition
-TestDatapath <- paste(exdir,"testclipsbinary",sep='')
-
-# Check folder names
-list.files(TestDatapath)
-
-# Create spectorgram images
-spectrogram_images(
-   trainingBasePath = TestDatapath,
-   outputBasePath = 'data/testimagesbinary/',
-   minfreq.khz = 0.4,
-   maxfreq.khz = 1.6,
-   splits = c(0, 0, 1), # Assign proportion to training, validation, or test folders
-   new.sampleratehz = 'NA'
- )
-
-```
-
-# Model training
+# Binary model training
 We will now train a binary model. We will use the spectrogram images for training, test and validation that we created above. Note that for "input.data.path" the train and valid folders need to be there. For the test data, the path must contain the 'test' folder. You can specify multiple model architectures and number of epochs for training.  
 
 ```{r eval = FALSE}
