### 2025-02-10 DenaJGibbon/gibbonNetR#6 @DenaJGibbon
I am closing this issue now, as comments were addressed and/or the examples were moved to the pkgdown site.
