+++ README.Rmd
@@ -17,11 +17,13 @@ knitr::opts_knit$set(root.dir = '/Users/denaclink/Desktop/RStudioProjects/gibbon
 
 # Overview
 
-This README provides code for training, testing, and deploying, different convolutional neural network model architectures on acoustic data using spectrogram images. Using this package you can compare different model architectures. Once you determine the best performing model you can deploy it over multiple sound files. You can also use a trained model as a 'feature extractor', and combined with an unsupervised clustering algorithm you can visualize differences in acoustic signals.  
+This README provides code for training, testing, and deploying, different convolutional neural network model architectures for automated detection and classification of acoustic data.
+
+Users can train both binary and multi-class classification models on spectrogram images, and evaluate their performance on test datasets. The package includes tools for performance evaluation, allowing easy identification of the best-performing model. Once the best-performing model is identified, it can be deployed for large-scale inference on multiple sound files. In addition to classification, trained CNNs can be used as feature extractors. Combined with unsupervised clustering, this enables visualization of differences in acoustic signals.
 
 # Usage
 
-A detailed usage guide can be found at: https://denajgibbon.github.io/gibbonNetR/
+A detailed usage guide can be found at: <https://denajgibbon.github.io/gibbonNetR/>
 
 # Installation
 
@@ -69,11 +71,8 @@ library(purrr)
     epoch.iterations = 1,  # Or any other list of integer epochs
     early.stop = "yes",
     save.model= FALSE,
-    output.base.path = paste(tempdir(),'/MultiDir/',sep=''),
+    output.base.path = file.path(tempdir(),'/MultiDir/',sep=''),
     trainingfolder = "test_multi",
     noise.category = 'noise'
   )
 ```
-
-
- 
