+++ vignettes/a01b-GettingStarted.Rmd
@@ -12,16 +12,19 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-# Getting started with your own data.
+## Getting started with your own data.
 You will need to have labeled .wav files that contain at least two classes, and one of them should be the 'noise' class. All .wav files for each class should be in a folder with the relevant class name.
 
-# Getting started multi-class example data.
+## Getting started multi-class example data.
 To follow best practices we want to use training and test data from different passive acoustic monitoring arrays. These clips are from Danum Valley Conservation Area, Sabah, Malaysia, but are taken from two different arrays. The "spectrogram_images" function takes folders organized and labeled by signal type, and creates training, validation, or test folders using the "splits" specified in the function call. In the example below 70% of the data is used for training and 30% for validation.   
 
-## We create spectrogram images of the sound files. 
+### We create spectrogram images of the sound files. 
 ```{r eval=FALSE}
 # Link to training clips on Zenodo
 ZenodoLink <- 'https://zenodo.org/records/14213067/files/trainingclips.zip?download=1'
+
+# Create data folder in your working directory
+dir.create("data")
 
 # Download into specified zip file location
 download.file(url = ZenodoLink, destfile = 'data/trainingclips.zip',method='curl')
@@ -53,7 +56,7 @@ spectrogram_images(
   <img src="spectro.png" align="center" width = "50%" height="50%"/>
   <figcaption>"Representative spectrogram images"</em>
 
-## Download example test files on Zenodo and convert to spectrogram images
+### Download example test files on Zenodo and convert to spectrogram images
 The "spectrogram_images" function below puts 100% of the test images into the test folder. 
 ```{r eval = FALSE}
 library(gibbonNetR)
@@ -86,13 +89,16 @@ spectrogram_images(
 
 ```
 
-# Get started with binary example data.
+## Get started with binary example data.
 
-## We create spectrogram images of the binary sound files. 
+### We create spectrogram images of the binary sound files. 
 
 ```{r eval=FALSE}
 # Link to training clips on Zenodo
 ZenodoLink <- 'https://zenodo.org/records/14213067/files/trainingclipsbinary.zip?download=1'
+
+# Create data folder in your working directory
+dir.create("data")
 
 # Download into specified zip file location
 download.file(url = ZenodoLink, destfile = 'data/trainingclipsbinary.zip',method='curl')
@@ -121,7 +127,7 @@ spectrogram_images(
 
 ```
 
-## Download example test files on Zenodo and convert to spectrogram images
+### Download example test files on Zenodo and convert to spectrogram images
 The "spectrogram_images" function below puts 100% of the test images into the test folder. 
 ```{r eval = FALSE}
 library(gibbonNetR)
