+++ pkgdown.yml
@@ -1,15 +1,14 @@
 pandoc: 3.1.11
-pkgdown: 2.0.9
+pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
   a01a-Overview: a01a-Overview.html
   a01b-GettingStarted: a01b-GettingStarted.html
-  a02-Training multiclass models: a02-Training multiclass models.html
-  a03-Training binary models: a03-Training binary models.html
-  a04-Automated detection: a04-Automated detection.html
+  a02-Trainingmulticlassmodels: a02-Trainingmulticlassmodels.html
+  a03-Trainingbinarymodels: a03-Trainingbinarymodels.html
+  a04-Automateddetection: a04-Automateddetection.html
   a05-UnsupervisedClustering: a05-UnsupervisedClustering.html
-last_built: 2025-02-10T15:37Z
+last_built: 2025-02-10T17:17Z
 urls:
   reference: https://denajgibbon.github.io/gibbonNetR/reference
   article: https://denajgibbon.github.io/gibbonNetR/articles
-
