+++ R/evaluate_trainedmodel_performance.R
@@ -32,8 +32,9 @@
 #' head(read.csv(CSVName[1]))
 #'
 #'}
-
+#' @importFrom utils write.csv read.csv
 #' @export
+#'
 evaluate_trainedmodel_performance <-
   function(trained_models_dir,
            image_data_dir,
