+++ R/deploy_CNN_multi.R
@@ -59,10 +59,17 @@
 #' )
 #'
 #' }
-
 #' @importFrom grDevices jpeg graphics.off
 #' @importFrom stats predict median
-#' @importFrom utils write.table write.csv read.csv
+#' @importFrom magrittr %>%
+#' @importFrom utils write.csv read.csv write.table
+#' @importFrom tuneR readWave writeWave
+#' @importFrom stringr str_detect str_count str_split_fixed
+#' @importFrom seewave spectro cutw
+#' @importFrom torch torch_tensor as_array torch_sigmoid
+#' @importFrom torchvision transform_to_tensor transform_color_jitter transform_resize transform_center_crop transform_normalize
+#' @importFrom luz luz_load
+#' @importFrom tools file_path_sans_ext
 #'
 #' @export
 
