+++ R/get_best_performance.R
@@ -95,110 +95,82 @@ get_best_performance <- function(performancetables.dir,
                                  class = 'hornbill.helmeted',
                                  Thresh.val = 0.5) {
   # Read all CSV files from the directory
-  FrozenFiles <-
-    list.files(performancetables.dir, full.names = TRUE)
+  FrozenFiles <- list.files(performancetables.dir, full.names = TRUE)
   FrozenCombined <- suppressMessages(map_dfr(FrozenFiles, read_csv))
 
   if (model.type == 'multi') {
-    if (class %in% FrozenCombined$Class == FALSE) {
+    if (!class %in% FrozenCombined[["Class"]]) {
       print(paste(
         'Not detected',
         class,
         'Here are the present classes:',
-        unique(FrozenCombined$Class)
+        unique(FrozenCombined[["Class"]])
       ))
-      break
+      return(NULL)  # Use return() instead of break
     }
-
-
 
     print(paste(
       'Evaluating performance for',
       class,
       'Here are the present classes:',
-      paste(unique(FrozenCombined$Class))
+      paste(unique(FrozenCombined[["Class"]]))
     ))
 
-    FrozenCombined <-
-      droplevels(subset(FrozenCombined, Class == class))
+    FrozenCombined <- droplevels(FrozenCombined[FrozenCombined[["Class"]] == class, ])
   }
 
-  unique_training_data <- unique(FrozenCombined$`Training Data`)
+  unique_training_data <- unique(FrozenCombined[["Training Data"]])
 
   best_f1_results <- data.frame()
   best_precision_results <- data.frame()
   best_recall_results <- data.frame()
   best_auc_results <- data.frame()
 
-  FrozenCombined <- subset(FrozenCombined, Threshold >= Thresh.val)
+  FrozenCombined <- FrozenCombined[FrozenCombined[["Threshold"]] >= Thresh.val, ]
 
-  # Loop through each 'TrainingData' type and extract best performance metrics
+  # Loop through each 'Training Data' type and extract best performance metrics
   for (td in unique_training_data) {
-    subset_data <- subset(FrozenCombined, `Training Data` == td)
-    MaxF1 <- max(subset_data$F1, na.rm = TRUE)
-    max_f1_row <- subset(subset_data, F1 == MaxF1)
+    subset_data <- FrozenCombined[FrozenCombined[["Training Data"]] == td, ]
+
+    max_f1_row <- subset_data[which.max(subset_data[["F1"]]), ]
     best_f1_results <- rbind(best_f1_results, max_f1_row)
 
-    MaxPrecision <- max(subset_data$Precision, na.rm = TRUE)
-    max_precision_row <-
-      subset(subset_data, Precision == MaxPrecision)
-    best_precision_results <-
-      rbind(best_precision_results, max_precision_row[which.max(max_precision_row$F1), ])
+    max_precision_row <- subset_data[which.max(subset_data[["Precision"]]), ]
+    best_precision_results <- rbind(best_precision_results, max_precision_row)
 
-    MaxRecall <-  max(subset_data$Recall, na.rm = TRUE)
-    max_recall_row <- subset(subset_data, Recall == MaxRecall)
-    best_recall_results <-
-      rbind(best_recall_results, max_recall_row[which.max(max_recall_row$F1), ])
+    max_recall_row <- subset_data[which.max(subset_data[["Recall"]]), ]
+    best_recall_results <- rbind(best_recall_results, max_recall_row)
 
-    max_auc_row <- subset_data[which.max(subset_data$AUC), ]
[…]
