+++ NAMESPACE
@@ -27,18 +27,32 @@ importFrom(ggpubr,ggline)
 importFrom(ggpubr,ggscatter)
 importFrom(grDevices,graphics.off)
 importFrom(grDevices,jpeg)
+importFrom(luz,luz_load)
 importFrom(magrittr,"%>%")
 importFrom(purrr,"%>%")
 importFrom(purrr,map_dfr)
 importFrom(readr,read_csv)
 importFrom(readr,write_csv)
+importFrom(seewave,cutw)
 importFrom(seewave,spectro)
 importFrom(stats,median)
 importFrom(stats,predict)
+importFrom(stringr,str_count)
 importFrom(stringr,str_detect)
 importFrom(stringr,str_replace)
 importFrom(stringr,str_split_fixed)
 importFrom(tibble,tibble)
 importFrom(tools,file_path_sans_ext)
+importFrom(torch,as_array)
+importFrom(torch,torch_sigmoid)
+importFrom(torch,torch_tensor)
+importFrom(torchvision,transform_center_crop)
+importFrom(torchvision,transform_color_jitter)
+importFrom(torchvision,transform_normalize)
+importFrom(torchvision,transform_resize)
+importFrom(torchvision,transform_to_tensor)
 importFrom(tuneR,readWave)
+importFrom(tuneR,writeWave)
+importFrom(utils,read.csv)
+importFrom(utils,write.csv)
 importFrom(utils,write.table)
