+++ pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   a03-Trainingbinarymodels: a03-Trainingbinarymodels.html
   a04-Automateddetection: a04-Automateddetection.html
   a05-UnsupervisedClustering: a05-UnsupervisedClustering.html
-last_built: 2025-02-12T12:41Z
+last_built: 2025-02-12T14:09Z
 urls:
   reference: https://denajgibbon.github.io/gibbonNetR/reference
   article: https://denajgibbon.github.io/gibbonNetR/articles
