+++ R/deploy_CNN_multi.R
@@ -113,7 +113,7 @@ deploy_CNN_multi <- function(output_folder,
     path_to_files_long <- list()
 
     for (a in 1:length(detect_pattern)) {
-      print(paste(
+      message(paste(
         'identifying sound files with the following pattern',
         detect_pattern[a]
       ))
@@ -136,8 +136,8 @@ deploy_CNN_multi <- function(output_folder,
       RavenSelectionTableDFTopModel <- data.frame()
 
       start.time.detection <- Sys.time()
-      print(paste(x, 'out of', length(path_to_files_long)))
-      print(path_to_files_short[x])
+      message(paste(x, 'out of', length(path_to_files_long)))
+      message(path_to_files_short[x])
       TempWav <- readWave(path_to_files_long[x])
       WavDur <- duration(TempWav)
 
@@ -146,7 +146,7 @@ deploy_CNN_multi <- function(output_folder,
 
       i <- 1
       while (i + clip_duration < WavDur) {
-        # print(i)
+        # message(i)
         Seq.start[[i]] = i
         Seq.end[[i]] = i + clip_duration
         i = i + hop_size
@@ -160,7 +160,7 @@ deploy_CNN_multi <- function(output_folder,
 
 
       # Subset sound clips for classification -----------------------------------
-      print('saving sound clips')
+      message('saving sound clips')
       set.seed(13)
       length <- nrow(TempClips)
 
@@ -201,7 +201,7 @@ deploy_CNN_multi <- function(output_folder,
                                       ))
 
         if (downsample_rate != 'NA') {
-          print('downsampling')
+          message('downsampling')
           short.sound.files <- lapply(1:length(short.sound.files),
                                       function(i)
                                         downsample(short.sound.files[[i]],
@@ -209,7 +209,7 @@ deploy_CNN_multi <- function(output_folder,
         }
 
         for (d in 1:length(short.sound.files)) {
-          #print(d)
+          #message(d)
           writeWave(
             short.sound.files[[d]],
             paste(
@@ -226,7 +226,7 @@ deploy_CNN_multi <- function(output_folder,
         }
 
         # Save images to a temp folder
-        print(paste('Creating images', start.time[1], 'start time clips'))
+        message(paste('Creating images', start.time[1], 'start time clips'))
 
         for (e in 1:length(short.sound.files)) {
           jpeg(
@@ -260,7 +260,7 @@ deploy_CNN_multi <- function(output_folder,
         }
 
         # Predict using TopModel ----------------------------------------------------
-        print('Classifying images using Top Model')
+        message('Classifying images using Top Model')
 
         test.input <- paste(tempdir(), '/Images/', sep = '')
 
@@ -297,7 +297,7 @@ deploy_CNN_multi <- function(output_folder,
 
         # Convert to a factor
         modelMultiPred <- as.factor(PredMPS)
-        print(modelMultiPred)
+        message(modelMultiPred)
 
         # Calculate the probability associated with each class
         Probability <-
@@ -333,7 +333,7 @@ deploy_CNN_multi <- function(output_folder,
         image.files.short <-
           str_split_fixed(image.files.short, pattern = '.jpg', n = 2)[, 1]
 
-        print('Saving output')
+        message('Saving output')
 
         Detections <-
           which(
@@ -382,7 +382,7 @@ deploy_CNN_multi <- function(output_folder,
             rep(single_class_category, length(DetectionIndices))
         }
 
-        print(paste(
+        message(paste(
           'Saving output to',
           paste(
             output_folder,
@@ -429,7 +429,7 @@ deploy_CNN_multi <- function(output_folder,
 
         Detections <- image.files.short[DetectionIndices]
[…]
