+++ NAMESPACE
@@ -23,6 +23,9 @@ import(torchvision)
 import(tuneR)
 import(umap)
 import(viridis)
+importFrom(ROCR,performance)
+importFrom(ROCR,prediction)
+importFrom(coro,loop)
 importFrom(ggpubr,ggline)
 importFrom(ggpubr,ggscatter)
 importFrom(grDevices,graphics.off)

[…]
