+++ pkgdown.yml
@@ -1,5 +1,5 @@
 pandoc: 3.1.11
-pkgdown: 2.0.9
+pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
   a01a-Overview: a01a-Overview.html
@@ -8,8 +8,7 @@ articles:
   a03-Trainingbinarymodels: a03-Trainingbinarymodels.html
   a04-Automateddetection: a04-Automateddetection.html
   a05-UnsupervisedClustering: a05-UnsupervisedClustering.html
-last_built: 2025-02-12T14:11Z
+last_built: 2025-02-12T15:06Z
 urls:
   reference: https://denajgibbon.github.io/gibbonNetR/reference
   article: https://denajgibbon.github.io/gibbonNetR/articles
-
