+++ R/gibbonNetR-package.R
@@ -4,3 +4,6 @@
 ## usethis namespace: start
 ## usethis namespace: end
 NULL
+
+## quiets concerns of R CMD check re: the .'s that appear in pipelines
+if(getRversion() >= "4.4.0")  utils::globalVariables(c(".", "self"))
