+++ NAMESPACE
@@ -12,6 +12,7 @@ export(train_CNN_multi)
 import(aricode)
 import(caret)
 import(cowplot)
+import(data.table)
 import(dbscan)
 import(ggplot2)
 import(luz)

[…]
