### 2025-02-12 7f02dce
Remove 'break' based on package check.
