+++ R/evaluate_trainedmodel_performance_multi.R
@@ -56,8 +56,9 @@ evaluate_trainedmodel_performance_multi <-
 
     if (length(trained_models) == 0) {
       message('No models in specified directory')
-      break
+      stop('Stopping execution: No models found.')  # Stops the execution
     }
+
 
     # List image files
     image_files <-
