### 2025-02-12 DenaJGibbon/gibbonNetR#9 @DenaJGibbon
Closing this as errors and warnings are gone. Examples seem to be running fine on my machine, but I will double-check them while updating the documentation.
