+++ paper/references.bib
@@ -6,7 +6,7 @@
 	date = {2023},
 	journal = {Frontiers in Ecology and Evolution},
 	volume = {11},
-	url = {https://www.frontiersin.org/articles/10.3389/fevo.2023.1071640}
+	doi = {10.3389/fevo.2023.1071640}
 }
 
 @article{sugai2019,
@@ -44,7 +44,8 @@
 	pages = {177{\textendash}196},
 	volume = {25},
 	number = {2},
-	note = {Publisher: Taylor & Francis}
+	note = {Publisher: Taylor & Francis},
+	url= {10.1080/09524622.2015.1133320}
 }
 
 @article{balantic2020,
@@ -117,7 +118,7 @@ Publisher: Nature Publishing Group},
 	author = {Chollet, {François} and others, },
 	year = {2015},
 	date = {2015},
-	url = {https://keras.io}
+	url = {https://github.com/fchollet/keras}
 }
 
 @inbook{paszke2019,
@@ -178,7 +179,8 @@ Publisher: Nature Publishing Group},
 	year = {2009},
 	date = {2009},
 	publisher = {Ieee},
-	pages = {248{\textendash}255}
+	pages = {248{\textendash}255},
+	url = {10.1109/cvpr.2009.5206848}
 }
 
 @article{dufourq2022,
@@ -189,7 +191,7 @@ Publisher: Nature Publishing Group},
 	journal = {Ecological Informatics},
 	pages = {101688},
 	volume = {70},
-	doi = {https://doi.org/10.1016/j.ecoinf.2022.101688},
+	doi = {10.1016/j.ecoinf.2022.101688},
 	url = {https://www.sciencedirect.com/science/article/pii/S1574954122001388}
 }
 
@@ -201,7 +203,7 @@ Publisher: Nature Publishing Group},
 	journal = {Applied Acoustics},
 	pages = {108939},
 	volume = {198},
-	doi = {https://doi.org/10.1016/j.apacoust.2022.108939},
+	doi = {10.1016/j.apacoust.2022.108939},
 	url = {https://www.sciencedirect.com/science/article/pii/S0003682X22003139}
 }
 
@@ -231,7 +233,8 @@ Publisher: Nature Publishing Group},
 	author = {Keydana, Sigrid},
 	year = {2023},
 	date = {2023},
-	publisher = {CRC Press}
+	publisher = {CRC Press},
+	url = {10.1201/9781003275923}
 }
 
 @article{lecun1995,
@@ -266,7 +269,8 @@ Publisher: Nature Publishing Group},
 	pages = {84{\textendash}90},
 	volume = {60},
 	number = {6},
-	note = {Publisher: AcM New York, NY, USA}
+	note = {Publisher: AcM New York, NY, USA},
+	url ={10.1145/3065386}
 }
 
 @article{simonyan2014,
@@ -282,7 +286,8 @@ Publisher: Nature Publishing Group},
 	author = {He, Kaiming and Zhang, Xiangyu and Ren, Shaoqing and Sun, Jian},
 	year = {2016},
 	date = {2016},
-	pages = {770{\textendash}778}
+	pages = {770{\textendash}778},
+	url ={10.1109/cvpr.2016.90}
 }
 
 @article{lakdari2024mel,
@@ -292,7 +297,8 @@ Publisher: Nature Publishing Group},
   volume={80},
   pages={102457},
   year={2024},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  url ={10.1016/j.ecoinf.2023.102457}
 }
 
 @article{sethi2020characterizing,
@@ -315,7 +321,8 @@ Publisher: Nature Publishing Group},
 	pages = {22876},
 	volume = {13},
 	number = {1},
[…]
