+++ paper/references.bib
@@ -492,5 +492,5 @@ abstract = {Passive acoustic monitoring – an approach that utilizes autonomous
   journal={Ecological Indicators},
   year={2024},
   note={Forthcoming},
-  url={https://www.elsevier.com/locate/ecolind},
+  url={10.1016/j.ecolind.2024.112501},
 }
