+++ tests/testthat/test-deploy_CNN_binary.R
@@ -35,6 +35,15 @@ test_that("Function outputs expected .csv file", {
 
   ListSelections <-  list.files(paste(tempdir(),'/BinaryDir/Results/Selections/',sep=''),full.names = T)
 
-  expect_true( ncol( read.delim(ListSelections[1])) ==10)
+  results <- ListSelections[1]
+
+  expect_true( ncol( read.delim(results)) ==10)
+
+  # Check that output file is created
+  expect_true(file.exists(results))
+
+  # Check that the data frame has expected columns
+  expect_true(all(c("Probability", "Detections", "Class") %in% colnames(read.delim(results))))
+
 
 })
