+++ R/deploy_CNN_multi.R
@@ -24,40 +24,39 @@
 #'
 #' @examples
 #' {
-#' # Load data
-#' data("TempBinWav")
+#'   # Load data
+#'   data("TempBinWav")
 #'
-#' # Create necessary directories
-#' dir.create(file.path(tempdir(), "MultiDir", "Wav"), recursive = TRUE, showWarnings = FALSE)
+#'   # Create necessary directories
+#'   dir.create(file.path(tempdir(), "MultiDir", "Wav"), recursive = TRUE, showWarnings = FALSE)
 #'
-#' # Write to temp directory
-#' writeWave(TempBinWav, filename = file.path(tempdir(), "MultiDir", "Wav", "TempBinWav.wav"))
+#'   # Write to temp directory
+#'   writeWave(TempBinWav, filename = file.path(tempdir(), "MultiDir", "Wav", "TempBinWav.wav"))
 #'
-#' # Set model directory
-#' trained_models_dir <- system.file("extdata", "trainedresnetmulti", package = "gibbonNetR")
+#'   # Set model directory
+#'   trained_models_dir <- system.file("extdata", "trainedresnetmulti", package = "gibbonNetR")
 #'
-#' # Specify model path
-#' ModelPath <- list.files(trained_models_dir, full.names = TRUE)
+#'   # Specify model path
+#'   ModelPath <- list.files(trained_models_dir, full.names = TRUE)
 #'
-#' # Deploy trained model over sound files
-#' deploy_CNN_multi(
-#'   clip_duration = 12,
-#'   architecture = "resnet18",
-#'   output_folder = file.path(tempdir(), "MultiDir", "Results", "Images"),
-#'   output_folder_selections = file.path(tempdir(), "MultiDir", "Results", "Selections"),
-#'   output_folder_wav = file.path(tempdir(), "MultiDir", "Results", "Wavs"),
-#'   detect_pattern = NA,
-#'   top_model_path = ModelPath,
-#'   path_to_files = file.path(tempdir(), "MultiDir", "Wav"),
-#'   downsample_rate = "NA",
-#'   save_wav = FALSE,
-#'   class_names = c("female.gibbon", "hornbill.helmeted", "hornbill.rhino", "long.argus", "noise"),
-#'   noise_category = "noise",
-#'   single_class = FALSE,
-#'   threshold = 0.25,
-#'   max_freq_khz = 2
-#' )
-#'
+#'   # Deploy trained model over sound files
+#'   deploy_CNN_multi(
+#'     clip_duration = 12,
+#'     architecture = "resnet18",
+#'     output_folder = file.path(tempdir(), "MultiDir", "Results", "Images"),
+#'     output_folder_selections = file.path(tempdir(), "MultiDir", "Results", "Selections"),
+#'     output_folder_wav = file.path(tempdir(), "MultiDir", "Results", "Wavs"),
+#'     detect_pattern = NA,
+#'     top_model_path = ModelPath,
+#'     path_to_files = file.path(tempdir(), "MultiDir", "Wav"),
+#'     downsample_rate = "NA",
+#'     save_wav = FALSE,
+#'     class_names = c("female.gibbon", "hornbill.helmeted", "hornbill.rhino", "long.argus", "noise"),
+#'     noise_category = "noise",
+#'     single_class = FALSE,
+#'     threshold = 0.25,
+#'     max_freq_khz = 2
+#'   )
 #' }
 #' @importFrom grDevices jpeg graphics.off
 #' @importFrom stats predict median
@@ -85,26 +84,31 @@ deploy_CNN_multi <- function(output_folder,
                              downsample_rate = 16000,
                              threshold = 0.1,
                              save_wav = TRUE,
-                             class_names = c('female.gibbon',
-                                             'hornbill.helmeted',
[…]
