+++ R/train_CNN_binary.R
@@ -25,18 +25,19 @@
 #' }
 #'
 #' @examples
-#' { input.data.path <- system.file("extdata", "binary/", package = "gibbonNetR")
+#' {
+#'   input.data.path <- system.file("extdata", "binary/", package = "gibbonNetR")
 #'   test.data <- system.file("extdata", "binary/test/", package = "gibbonNetR")
 #'   result <- train_CNN_binary(
 #'     input.data.path = input.data.path,
 #'     test.data = test.data,
-#'     architecture = "alexnet",  # Choose architecture
+#'     architecture = "alexnet", # Choose architecture
 #'     unfreeze.param = TRUE,
 #'     batch_size = 6,
 #'     learning_rate = 0.001,
-#'     epoch.iterations = 1,  # Or any other list of integer epochs
+#'     epoch.iterations = 1, # Or any other list of integer epochs
 #'     early.stop = "yes",
-#'     output.base.path = paste(tempdir(),'/',sep=''),
+#'     output.base.path = paste(tempdir(), "/", sep = ""),
 #'     trainingfolder = "test_binary"
 #'   )
 #'   print(result)
@@ -62,17 +63,18 @@ train_CNN_binary <-
            learning_rate,
            save.model = FALSE,
            epoch.iterations = 1,
-           early.stop = 'yes',
-           output.base.path = 'data/',
+           early.stop = "yes",
+           output.base.path = "data/",
            trainingfolder,
            list.thresholds = seq(0.1, 1, .1),
            positive.class = "Gibbons",
            negative.class = "Noise") {
     # Device
-    device <- if (cuda_is_available())
+    device <- if (cuda_is_available()) {
       "cuda"
-    else
+    } else {
       "cpu"
+    }
 
     to_device <- function(x, device) {
       x$to(device = device)
@@ -81,17 +83,19 @@ train_CNN_binary <-
     # Location to save the output
     output.data.path <-
       paste(output.base.path,
-            trainingfolder,
-            'binary',
-            'unfrozen',
-            unfreeze.param,
-            '/',
-            sep = '_')
+        trainingfolder,
+        "binary",
+        "unfrozen",
+        unfreeze.param,
+        "/",
+        sep = "_"
+      )
 
     # Create if doesn't exist
     dir.create(output.data.path,
-               showWarnings = FALSE,
-               recursive = TRUE)
+      showWarnings = FALSE,
+      recursive = TRUE
+    )
 
     # Metadata
     metadata <- data.frame(
@@ -108,18 +112,20 @@ train_CNN_binary <-
       Negative.class = negative.class
     )
 
-    write.csv(metadata,
-              paste0(output.data.path, architecture, "_model_metadata.csv"))
+    write.csv(
+      metadata,
+      paste0(output.data.path, architecture, "_model_metadata.csv")
+    )
 
     for (a in 1:length(epoch.iterations)) {
-      message(paste('Training', architecture))
-      n.epoch <- epoch.iterations [a]
+      message(paste("Training", architecture))
+      n.epoch <- epoch.iterations[a]
 
 
       if (architecture %in% c("alexnet", "vgg16", "vgg19") == TRUE) {
         # Data loaders setup
         train_ds <- image_folder_dataset(
-          file.path(input.data.path, 'train'),
+          file.path(input.data.path, "train"),
           transform = . %>%
             torchvision::transform_to_tensor() %>%
             torchvision::transform_resize(size = c(224, 224)) %>%
@@ -128,8 +134,9 @@ train_CNN_binary <-
               mean = c(0.485, 0.456, 0.406),
               std = c(0.229, 0.224, 0.225)
             ),
-          target_transform = function(x)
+          target_transform = function(x) {
             as.double(x) - 1
+          }
         )
 
         valid_ds <- image_folder_dataset(
[…]
