+++ R/gibbonNetR-package.R
@@ -6,4 +6,4 @@
 NULL
 
 ## quiets concerns of R CMD check re: the .'s that appear in pipelines
-if(getRversion() >= "4.4.0")  utils::globalVariables(c(".", "self"))
+if (getRversion() >= "4.4.0") utils::globalVariables(c(".", "self"))

[…]
