+++ R/spectrogram_images.R
@@ -82,7 +82,9 @@
 #'   )
 #'
 #'   # Create a dataloader
-#'   train_dl <- dataloader(train_ds, batch_size = train_ds$.length(), shuffle = FALSE, drop_last = TRUE)
+#'   train_dl <- dataloader(train_ds,
+#'   batch_size = train_ds$.length(),
+#'   shuffle = FALSE, drop_last = TRUE)
 #'
 #'   # Extract the next batch from the dataloader
 #'   batch <- train_dl$.iter()$.next()
