### 2025-02-14 DenaJGibbon/gibbonNetR#7 @DenaJGibbon
I ran pkgcheck and applied the 'styler' package to format the code.
