+++ R/deploy_CNN_binary.R
@@ -17,7 +17,7 @@
 #' @param positive.class A character string specifying the positive class label. Default is 'Gibbons'.
 #' @param negative.class A character string specifying the negative class label. Default is 'Noise'.
 #' @param max_freq_khz The maximum frequency in kHz for spectrogram visualization. Default is 2.
-
+#' @return Returns spectrogram images, wav files (if specified), and Raven selection tables for each sound file.
 #' @details This function processes sound data from a directory, extracts sound clips, converts them to images, performs image classification using a pre-trained deep learning model, and saves the results including selection tables and image and audio files.
 #'
 #' @examples
