+++ DESCRIPTION
@@ -47,3 +47,4 @@ Config/testthat/edition: 3
 VignetteBuilder: knitr
 LazyData: true
 URL: https://denajgibbon.github.io/gibbonNetR/
+BugReports: https://github.com/DenaJGibbon/gibbonNetR/issues
