+++ vignettes/a03-Trainingbinarymodels.Rmd
@@ -29,7 +29,7 @@ options(width = 100, max.print = 100)
 ```
 
 # Binary model training
-We will now train a binary model. We will use the spectrogram images for training, test and validation that we created above. Note that for "input.data.path" the train and valid folders need to be there. For the test data, the path must contain the 'test' folder. You can specify multiple model architectures and number of epochs for training.  
+We will now train a binary model. We will use the spectrogram images for training, test and validation that we created previously Note that for "input.data.path" the train and valid folders need to be there. For the test data, the path must contain the 'test' folder. You can specify multiple model architectures and number of epochs for training.  
 
 ```{r eval = FALSE}
 # Location of spectrogram images for training

[…]
