### 2025-02-15 DenaJGibbon/gibbonNetR#5 @DenaJGibbon
Descriptions of data added to pkgdown site and link to Zenodo site: https://denajgibbon.github.io/gibbonNetR/articles/a01b-GettingStarted.html.
