### 2025-02-15 DenaJGibbon/gibbonNetR#5 @DenaJGibbon
Added notes in the function documentation indicating required inputs from other functions.
