### 2025-02-15 f59fa6c
Update function documentation.
