+++ R/data.R
@@ -9,9 +9,9 @@
 #' \describe{
 #'   \item{country}{Malaysia}
 #'   \item{year}{2024}
-#'   ...
+#'   \item{Note}{See: https://zenodo.org/records/14213067 for more details}
 #' }
 
 
-#' @source <https://doi.org/10.5281/zenodo.10927637>
+#' @source <https://zenodo.org/records/14213067>
 "TempBinWav"
