+++ R/deploy_CNN_multi.R
@@ -22,6 +22,7 @@
 #' @param for_prrec Whether to output all detections to create a PR curve.
 #' @details This function processes sound data from a directory, extracts sound clips, converts them to images, performs image classification using a pre-trained deep learning model, and saves the results including selection tables and image and audio files.
 #' @return Returns spectrogram images, wav files (if specified), and Raven selection tables for each sound file.
+#' @note This function takes a model trained using the 'train_CNN_multi' function.
 #' @examples
 #' {
 #'   # Load data
