+++ vignettes/a01b-GettingStarted.Rmd
@@ -61,6 +61,8 @@ spectrogram_images(
 ```
 
 ## Getting started with multi-class example data.
+For more information on the example data included please refer to: https://zenodo.org/records/14213067.
+
 To follow best practices we want to use training and test data from different passive acoustic monitoring arrays. These clips are from Danum Valley Conservation Area, Sabah, Malaysia, but are taken from two different arrays. The "spectrogram_images" function takes folders organized and labeled by signal type, and creates training, validation, or test folders using the "splits" specified in the function call. In the example below 70% of the data is used for training and 30% for validation.   
 
 ### We create spectrogram images of the sound files. 
