+++ man/get_best_performance.Rd
@@ -28,13 +28,17 @@ and plots visualizing these metrics.
 Given the path to a directory of performance tables, this function reads in the tables,
 combines them, and extracts the best performance results based on various criteria.
 }
+\note{
+'train_CNN_binary' and 'train_CNN_multi' output a directory of 'performance tables'.
+This function requres that directory as input.
+}
 \examples{
 {
   # Simulate data for performance tables
   set.seed(123)
 
   #' Set directory
-  performance_tables_dir <- paste(tempdir(), "/example_performance_tables/", sep = "")
+  performance_tables_dir <- paste(tempdir(), "/performance_tables/", sep = "")
 
   #' Create directory for performance tables (NOTE THIS IS FOR TESTING ONLY)
   dir.create(performance_tables_dir, showWarnings = FALSE, recursive = TRUE)
