### 2025-02-15 DenaJGibbon/gibbonNetR#8 @DenaJGibbon
We are incredibly grateful for the detailed comments. We made the following modifications to the manuscript.
•	Updated the transfer learning section to clarify terms.
•	Added a paragraph to the state of the field highlighting existing packages in R.
•	Removed a few of the code examples and linked to the pkgdown site for detailed instructions on usage.
•	Added more detail to the embeddings section.
•	Updated the future directions to emphasize the positive aspects of the package.
