+++ paper/paper.md
@@ -16,11 +16,11 @@ affiliations:
  - name: Institute for Tropical Biology and Conservation,Universiti Malaysia Sabah (UMS), Kota Kinabalu, Sabah, Malaysia
    index: 2
 
-date: 2024
+date: 2025
 bibliography: references.bib
 ---
 # Summary
-Automated detection of acoustic signals is crucial for effective monitoring of vocal animals and their habitats across large spatial and temporal scales. Recent advances in deep learning have made high performance automated detection approaches more accessible to more practitioners. However, there are few deep learning approaches that can be implemented natively in R. The 'torch for R' ecosystem has made the use of convolutional neural networks (CNNs) accessible for R users. Here, we provide an R package and workflow to use CNNs for automated detection and classification of acoustics signals from passive acoustic monitoring data collected in Sabah, Malaysia. The package provides functions to create spectogram images from labeled data, compare the performance of different CNN architectures, deploy trained models over directories of sound files, and extract embeddings from trained models. The R programming language remains one of the most commonly used languages among ecologists, and we hope that this package makes deep learning approaches more accessible to this audience. In addition, these models can serve as important benchmarks for future automated detection work.  
+Automated detection of acoustic signals is crucial for effective monitoring of vocal animals and their habitats across large spatial and temporal scales. Recent advances in deep learning have made high performance automated detection approaches accessible to more practitioners. However, there are few deep learning approaches that can be implemented natively in R. The 'torch for R' ecosystem has made the use of convolutional neural networks (CNNs) accessible for R users. Here, we provide an R package and workflow to use CNNs for automated detection and classification of acoustics signals from passive acoustic monitoring data. We provide examples using data collected in Sabah, Malaysia. The package provides functions to create spectrogram images from labeled data, compare the performance of different CNN architectures, deploy trained models over directories of sound files, and extract embeddings from trained models. The R programming language remains one of the most commonly used languages among ecologists, and we hope that this package makes deep learning approaches more accessible to this audience. In addition, these models can serve as important benchmarks for future automated detection work.  
 
 # Statement of need
 
@@ -92,7 +92,8 @@ We can calculate the Normalize Mutual Information score, which provides a value 
 
 # Future directions
 
-There have been huge advances in the fields of deep learning and automated detection approaches for PAM data in recent years. The approach presented in this package is one of the first to use the 'torch for R' ecosystem, and employ automated detection using deep learning natively in R. More recent approaches use transfer learning from models that are explicitly trained on bioacoustics data, such as BirdNET [@ghani2023], have been introduced. There is a huge need in the field of bioacoustics to do benchmarking, wherein different model architectures and performance are compared across diverse datasets. The methods presented here can provide important benchmarks for future work, and for understanding how and if advances improve performance over more traditional methods. In addition, this package provides a comprehensive suite of tools for processing, analyzing, and visualizing acoustic data using deep learning, offering robust support for tasks such as automated signal detection, feature extraction, classification, and data visualization, which are critical in environmental and bioacoustic research. The R package is available on [Github](https://github.com/DenaJGibbon/gibbonNetR), where issues can be opened.
[…]
