### 2025-02-15 11a245f
Upgrade to torch 0.14.2 introduced breaking changes in extract_embeddings
