### 2025-02-15 6e5f7b8
Update extract_embeddings.
