### 2025-02-15 #7250 @DenaJGibbon
Hi everyone! I want to thank you all for your patience. I have been able to work through most of @steffilazerte's comments. They were so very helpful! I really appreciate it, and I think the package and documentation are much improved. As I was doing a final check I updated all the versions of the packages, and this caused a breaking change to one of the functions (extract_embeddings). I am still trouble-shooting how to fix this one issue. I will update as soon as possible. Thanks again!
