### 2025-02-15 DenaJGibbon/gibbonNetR#7 @DenaJGibbon
I moved the remaining 'pkgcheck' suggestions to another issue with remaining things to address.
