+++ R/spectrogram_images.R
@@ -82,7 +82,7 @@
 #'   )
 #'
 #'   # Create a dataloader
-#'   train_dl <- dataloader(train_ds,
+#'   train_dl <- torch::dataloader(train_ds,
 #'   batch_size = train_ds$.length(),
 #'   shuffle = FALSE, drop_last = TRUE)
 #'
@@ -95,7 +95,7 @@
 #'   class_names <- stringr::str_split_fixed(class_names, pattern = "/", n = 2)[, 1]
 #'
 #'   # Convert batch tensor of images to an array and reorder dimensions
-#'   images <- as_array(batch[[1]]) %>% aperm(perm = c(1, 3, 4, 2))
+#'   images <- torch::as_array(batch[[1]]) %>% aperm(perm = c(1, 3, 4, 2))
 #'
 #'   # Define a function to normalize pixel values
 #'   normalize_pixel_values <- function(image) {
@@ -121,6 +121,7 @@
 #' @importFrom seewave spectro
 #' @importFrom tools file_path_sans_ext
 #' @importFrom stringr str_split_fixed
+#' @importFrom torch dataloader as_array
 #' @export
 
 spectrogram_images <- function(trainingBasePath,
