+++ README.Rmd
@@ -43,6 +43,10 @@ install.packages("devtools")
 # Install gibbonNetR
 devtools::install_github("https://github.com/DenaJGibbon/gibbonNetR")
 
+# NOTE: As of February 15, 2025 need to install 
+# an older version of 'torch'
+remotes::install_version("torch", version = "0.13.0")
+
 # The first time you use the package 'torch' will need to install additional packages. You can start the process using the following:
 library(torch)
 ```
