+++ NAMESPACE
@@ -48,6 +48,7 @@ importFrom(stringr,str_split_fixed)
 importFrom(tibble,tibble)
 importFrom(tools,file_path_sans_ext)
 importFrom(torch,as_array)
+importFrom(torch,dataloader)
 importFrom(torch,torch_sigmoid)
 importFrom(torch,torch_tensor)
 importFrom(torchvision,transform_center_crop)
