### 2025-02-15 #7250 @DenaJGibbon
For now, to address the torch version issue I specified the required version required in the README, and also included a script that will return an error message if the necessary version is not installed. I would love any ideas on how to deal with this more effectively. Thanks so much!
