### 2025-02-18 #7250 @DenaJGibbon
Hello @faroit. I was wondering if there is a way to update the manuscript draft in the review system? Thanks!
