### 2025-02-22 2da377c
Add informative error message to spectrogram images function.
