### 2025-02-23 8485aed
Add option to change window length in settings
