+++ NAMESPACE
@@ -27,6 +27,7 @@ import(viridis)
 importFrom(ROCR,performance)
 importFrom(ROCR,prediction)
 importFrom(coro,loop)
+importFrom(dplyr,bind_rows)
 importFrom(ggpubr,ggline)
 importFrom(ggpubr,ggscatter)
 importFrom(grDevices,graphics.off)
@@ -34,6 +35,7 @@ importFrom(grDevices,jpeg)
 importFrom(luz,luz_load)
 importFrom(magrittr,"%>%")
 importFrom(purrr,"%>%")
+importFrom(purrr,map)
 importFrom(purrr,map_dfr)
 importFrom(readr,read_csv)
 importFrom(readr,write_csv)
