### 2025-02-23 dec751c
Update detections so that the have probability at the start to be comparable with BirdNET review tab
