### 2025-03-09 #7250 @faroit
> @faroit, I'm not sure if they require more explicit animal research statements as the package does include vocalisations recorded from animals?

@steffilazerte I am not familiar with animal research, should I check with the JOSS editors or some other experts if this is required in this case?
