### 2025-03-10 #7250 @Desjonqu
Hi again @faroit, sorry but I'm not sure where I can find my review checklist and where I'm supposed to post my review. I've just read again the reviewer guide (https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) but I can't quite find where and how to submit my review.
