### 2025-03-10 #7250 @DenaJGibbon
> > [@faroit](https://github.com/faroit), I'm not sure if they require more explicit animal research statements as the package does include vocalisations recorded from animals?
>
> [@steffilazerte](https://github.com/steffilazerte) I am not familiar with animal research, should I check with the JOSS editors or some other experts if this is required in this case?

We had all the required institutional approvals! I will add to the manuscript now.
