### 2025-03-11 #7250 @faroit
@Desjonqu generate your checklist as mentioned above: `@editorialbot generate my checklist
`
