### 2025-03-18 #7250 @Desjonqu
Dear @DenaJGibbon, I have completed my assessment of your paper and software. It took me a bit of time to reinstall the older version of torch but I made it work.

I find your paper really well written, comprehensive and clear. The package you developped is useful and fills a clear gap. I have a few minor suggestions to enable this package to be as user friendly as possible:
- Would it be possible to add a full working example showing the different functions, their arguments and how they work?
- Would you consider adding the complete scripts to obtain your paper's figures? I was not able to replicate the performance figure (figure 2) even with the code provided and I don't really now where to start to build figure 3.
- Could you also include a description of the output format?

Thanks again for this package which will definitely be useful to our community.
