### 2025-03-19 6fefaa6
Remove torch version dependency & update 'extract_embeddings'.
