+++ man/train_CNN_multi.Rd
@@ -16,6 +16,9 @@ train_CNN_multi(
   epoch.iterations = 1,
   early.stop = "yes",
   output.base.path = tempdir(),
+  brightness = 1,
+  contrast = 1,
+  saturation = 1,
   trainingfolder,
   noise.category = "Noise"
 )
