### 2025-03-24 ce20184
Update color jitter function.
