+++ R/evaluate_trainedmodel_performance.R
@@ -96,7 +96,6 @@ evaluate_trainedmodel_performance <-
       if (str_detect(model_type, pattern = "resnet")) {
         transform_list <- . %>%
           torchvision::transform_to_tensor() %>%
-          torchvision::transform_color_jitter() %>%
           transform_resize(256) %>%
           transform_center_crop(224) %>%
           transform_normalize(
