+++ man/train_CNN_binary.Rd
@@ -9,9 +9,9 @@ train_CNN_binary(
   test.data,
   architecture,
   noise.weight = 0.5,
-  brightness = 1,
-  contrast = 1,
-  saturation = 1,
+  brightness = 0,
+  contrast = 0,
+  saturation = 0,
   unfreeze.param = TRUE,
   batch_size = 32,
   learning_rate,
