+++ vignettes/a06-CompleteExample.Rmd
@@ -7,7 +7,10 @@ output:
     number_sections: true
     self_contained: yes
 vignette: >
-  %\VignetteIndexEntry{5. Unsupervised clustering}
+  %\VignetteIndexEntry{6. A complete working example}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
+
+
+
