### 2025-03-24 c5c2aa5
Fix error in 'get_best_performance'. Previous code was taking the first high metric, and ignoring cases where there were multiple.
