### 2025-04-01 a3b4c49
Fix AUC-ROC calculation for multiclass models.
