+++ NAMESPACE
@@ -48,6 +48,7 @@ importFrom(stringr,str_detect)
 importFrom(stringr,str_replace)
 importFrom(stringr,str_split_fixed)
 importFrom(tibble,tibble)
+importFrom(tidyr,pivot_longer)
 importFrom(tools,file_path_sans_ext)
 importFrom(torch,as_array)
 importFrom(torch,dataloader)
