+++ DESCRIPTION
@@ -35,6 +35,7 @@ Imports:
     stats,
     stringr,
     tibble,
+    tidyr,
     torchdatasets,
     umap,
     utils,
