+++ pkgdown.yml
@@ -9,7 +9,7 @@ articles:
   a04-Automateddetection: a04-Automateddetection.html
   a05-UnsupervisedClustering: a05-UnsupervisedClustering.html
   a06-CompleteExample: a06-CompleteExample.html
-last_built: 2025-04-02T12:34Z
+last_built: 2025-04-02T12:36Z
 urls:
   reference: https://denajgibbon.github.io/gibbonNetR/reference
   article: https://denajgibbon.github.io/gibbonNetR/articles
