+++ pkgdown.yml
@@ -10,7 +10,7 @@ articles:
   a05-UnsupervisedClustering: a05-UnsupervisedClustering.html
   a06-CompleteExample: a06-CompleteExample.html
   a07-JOSS figures: a07-JOSS figures.html
-last_built: 2025-04-02T14:10Z
+last_built: 2025-04-02T14:22Z
 urls:
   reference: https://denajgibbon.github.io/gibbonNetR/reference
   article: https://denajgibbon.github.io/gibbonNetR/articles
