+++ R/train_CNN_multi.R
@@ -16,11 +16,15 @@
 #' @param trainingfolder Character. A descriptive name for the training data, used for naming output files.
 #' @param noise.category Character. Label for the noise category. Default is "Noise".
 #'
-#' @return A list containing two elements:
+#' @return The function generates multiple output files, including:
 #' \itemize{
-#'   \item \strong{Output_Path}: Path where the trained model and metadata are saved.
-#'   \item \strong{Metadata}: A dataframe containing metadata about the training session.
+#'   \item \strong{Trained Models}: If save.model = TRUE outputs saved model files (.pt) for specified architectures.
+#'   \item \strong{Training Logs}: logs_model.csv containing logs of training sessions, including loss and accuracy metrics.
+#'   \item \strong{Metadata}: model_metadata.csv contains metadata from training run
+#'   \item \strong{Model predictions}: Saved for each architecture in output_TrainedModel_testdata.csv
+#'   \item \strong{Performance Evaluation}: Saves .csv summarizing performance for each architecture in nested folder in output.base.path 'performance_tables_multi'.
 #' }
+
 #'
 #' @examples
 #' {{ input.data.path <- system.file("extdata", "multiclass/", package = "gibbonNetR")
@@ -407,6 +411,17 @@ train_CNN_multi <- function(input.data.path, test.data, architecture,
     outputTableMulti <- cbind.data.frame(Probability,Folder)
     colnames(outputTableMulti) <- c(class_names,"ActualClass" )
 
+    # Save the output table as CSV file
+    write.csv(
+      outputTableMulti,
+      paste(
+        output.data.path,
+        trainingfolder, n.epoch, architecture,
+        "output_TrainedModel_testdata.csv",
+        sep = "_"
+      ),
+      row.names = FALSE
+    )
 
     UniqueClasses <- class_names
     UniqueClasses <- UniqueClasses[-which(UniqueClasses == noise.category)]
