+++ R/train_CNN_binary.R
@@ -18,10 +18,13 @@
 #' @param negative.class Character. The name of the negative class label. Default is 'Noise'.
 #' @param list.thresholds Numerical list indicating thresholds. Default is seq(0.1,1,.1).
 #'
-#' @return A list containing two elements:
+#' @return The function generates multiple output files, including:
 #' \itemize{
-#'   \item \strong{Output_Path}: The path where the model and metadata are saved.
-#'   \item \strong{Metadata}: A dataframe containing metadata about the training session.
+#'   \item \strong{Trained Models}: If save.model = TRUE outputs saved model files (.pt) for specified architectures.
+#'   \item \strong{Training Logs}: logs_model.csv containing logs of training sessions, including loss and accuracy metrics.
+#'   \item \strong{Metadata}: model_metadata.csv contains metadata from training run
+#'   \item \strong{Model predictions}: Saved for each architecture in output_TrainedModel_testdata.csv
+#'   \item \strong{Performance Evaluation}: Saves .csv summarizing performance for each architecture in nested folder in output.base.path 'performance_tables'.
 #' }
 #'
 #' @examples
@@ -547,14 +550,14 @@ train_CNN_binary <-
 
       outputTableTrainedModel$Probability <-
         1 - outputTableTrainedModel$Probability
-      # Save the output table as CSV file
+
+       # Save the output table as CSV file
       write.csv(
         outputTableTrainedModel,
         paste(
           output.data.path,
-          trainingfolder,
-          n.epoch,
-          "output_TrainedModel.csv",
+          trainingfolder, n.epoch, architecture,
+          "output_TrainedModel_testdata.csv",
           sep = "_"
         ),
         row.names = FALSE
