### 2025-04-02 #7250 @DenaJGibbon
Hello,

I have made updates to the package documentation based on the points highlighted by @Desjonqu.

_- Would it be possible to add a full working example showing the different functions, their arguments and how they work?_
There is a new page that includes a full working example: https://denajgibbon.github.io/gibbonNetR/articles/a06-CompleteExample.html

_- Would you consider adding the complete scripts to obtain your paper's figures? I was not able to replicate the performance figure (figure 2) even with the code provided and I don't really now where to start to build figure 3._
There is now a page in the documentation to recreate the figures. It may take a very long time to run the models depending on computing power, which is why all the other examples are for fewer epochs. @faroit please let me know if there is a more appropriate way to do this. https://denajgibbon.github.io/gibbonNetR/articles/a07-JOSS%20figures.html

_- Could you also include a description of the output format?_
 This has been updated for each of the functions that were missing.

I was also able to figure out the torch version issue, so the package now works with the most recent version of torch v0.14.2.
