### 2025-04-05 #7250 @faroit
> @faroit what are the suggested best practices by JOSS to include code to reproduce the figures in the paper? To get it work, I currently have the figures as images.

@DenaJGibbon sorry, I didn't see that message. I guess most JOSS authors use figshare to share reproducible figures but I don't think this would be very useful in your case. I find your documentation well done so I think this is sufficiently addressed. @Desjonqu do you agree?
