### 2025-04-05 8ed431b
Update AUC multi calculation. Previous way mixed probabilities of different classes.
