### 2025-04-08 #7250 @Desjonqu
> [@DenaJGibbon](https://github.com/DenaJGibbon) sorry, I didn't see that message. I guess most JOSS authors use figshare to share reproducible figures but I don't think this would be very useful in your case. I find your documentation well done so I think this is sufficiently addressed. [@Desjonqu](https://github.com/Desjonqu) do you agree?

It looks great as it is now. Thanks!
