### 2025-04-14 #7250 @faroit
@DenaJGibbon as I understand at this point the next steps are

1. There are two unchecked boxes in https://github.com/openjournals/joss-reviews/issues/7250#issuecomment-2714702910 to be reviewed by @Desjonqu
2. There are a large number of unchecked boxes to be reviewed by @steffilazerte (I guess almost all of them have been addressed)
3. There are is another separate review checklist from @steffilazerte to be reviewed and checked https://github.com/openjournals/joss-reviews/issues/7250#issuecomment-2432820127
4. There is one open-issue https://github.com/DenaJGibbon/gibbonNetR/issues/7 to be closed by @steffilazerte when she thinks this has been sufficiently addressed
5. when 1-4 are done, we can proceed with the post-review checklist

please let me know if you have more questions
