+++ tests/testthat/test-evaluate_trainedmodel_performance.R
@@ -16,7 +16,27 @@ test_that("function returns expected values", {
   # Find the location of saved evaluation files
   CSVName <- list.files(paste(tempdir(), "/data/", sep = ""), recursive = TRUE, full.names = TRUE)
 
-  # Check the output of the first file
-  head(read.csv(CSVName[1]))
-  expect_true(ncol(read.csv(CSVName[1])) == 19)
+  # List output files
+  CSVFiles <- list.files(paste(tempdir(), "/data/", sep = ""), recursive = TRUE, pattern = "\\.csv$", full.names = TRUE)
+  expect_true(length(CSVFiles) > 0)
+
+  # Read and check contents of first CSV
+  eval_df <- read.csv(CSVFiles[1])
+
+  # Check expected structure
+  expect_true(ncol(eval_df) == 19)
+  expect_true(all(c("AUC", "Class", "TestData", "Sensitivity", "Specificity", "F1") %in% colnames(eval_df)))
+
+  # Check number of rows > 0
+  expect_true(nrow(eval_df) > 0)
+
+  # Check AUC values between 0 and 1
+  expect_true(all(eval_df$AUC >= 0 & eval_df$AUC <= 1))
+
+  # Confirm that both classes are represented
+  expect_true(all(c("Gibbons") %in% eval_df$Class))
+
+  # Confirm accuracy metric values are in reasonable range
+  expect_true(all(eval_df$Sensitivity >= 0 & eval_df$Sensitivity <= 1))
+  expect_true(all(eval_df$Specificity >= 0 & eval_df$Specificity <= 1))
 })
