+++ NAMESPACE
@@ -42,6 +42,7 @@ importFrom(readr,write_csv)
 importFrom(seewave,cutw)
 importFrom(seewave,spectro)
 importFrom(stats,median)
+importFrom(stats,na.omit)
 importFrom(stats,predict)
 importFrom(stringr,str_count)
 importFrom(stringr,str_detect)
