### 2025-04-19 0aaf1fc
Remove duplicate figure legend numbers
