+++ _pkgdown.yml
@@ -1,4 +1,0 @@
-url: https://denajgibbon.github.io/gibbonNetR/
-template:
-  bootstrap: 5
-
