### 2025-04-19 1313ec8
Update figure legends.
