+++ pkgdown.yml
@@ -9,9 +9,6 @@ articles:
   a04-Automateddetection: a04-Automateddetection.html
   a05-UnsupervisedClustering: a05-UnsupervisedClustering.html
   a06-CompleteExample: a06-CompleteExample.html
-  a07-JOSS figures: a07-JOSS figures.html
-last_built: 2025-04-02T15:46Z
-urls:
-  reference: https://denajgibbon.github.io/gibbonNetR/reference
-  article: https://denajgibbon.github.io/gibbonNetR/articles
+  a07-JOSSfigures: a07-JOSSfigures.html
+last_built: 2025-04-19T14:57Z
 
