+++ paper/references.bib
@@ -7,6 +7,7 @@
     volume = {21},
     number = {20},
     pages = {7881},
+    doi = {10.1093/bioinformatics/bti623},
     url = {http://rocr.bioinf.mpi-sb.mpg.de},
   }
 
@@ -17,13 +18,15 @@
     journal = {Bioacoustics},
     volume = {18},
     pages = {213-226},
+    doi = {10.1080/09524622.2008.9753600}
   }
 
 @article{clink2019gibbonfindr,
   title={gibbonR: An R package for the detection and classification of acoustic signals},
   author={Clink, D. J. and Klinck, Holger},
   journal={arXiv preprint arXiv:1906.02572},
-  year={2019}
+  year={2019},
+  doi={https://doi.org/10.48550/arXiv.1906.02572}
 }
 
 @article{clinkzenodo2024,
@@ -48,7 +51,8 @@
   number={2},
   pages={184--191},
   year={2017},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={https://dx.doi.org/10.1111/2041-210X.12624}
 }
 
 @article{clink2023,
@@ -97,7 +101,7 @@
 	volume = {25},
 	number = {2},
 	note = {Publisher: Taylor & Francis},
-	url= {10.1080/09524622.2015.1133320}
+	doi= {10.1080/09524622.2015.1133320}
 }
 
 @article{balantic2020,
@@ -109,7 +113,8 @@
 	pages = {869{\textendash}877},
 	volume = {11},
 	number = {7},
-	note = {Publisher: Wiley Online Library}
+	note = {Publisher: Wiley Online Library},
+	doi={https://doi.org/10.1111/2041-210X.13397}
 }
 
 @article{kalan2015,
@@ -146,7 +151,8 @@ Publisher: Nature Publishing Group},
 	author = {Scavetta, Rick J and Angelov, Boyan},
 	year = {2021},
 	date = {2021},
-	publisher = {O'Reilly Media, Inc.}
+	publisher = {O'Reilly Media, Inc.},
+	doi={10.18637/jss.v103.b02}
 }
 
 @article{lawlor2022,
@@ -170,7 +176,8 @@ Publisher: Nature Publishing Group},
 	author = {Chollet, {François} and others, },
 	year = {2015},
 	date = {2015},
-	url = {https://github.com/fchollet/keras}
+	url = {https://github.com/fchollet/keras},
+	doi = {10.1163/1574-9347_bnp_e612900}
 }
 
 @inbook{paszke2019,
@@ -180,7 +187,8 @@ Publisher: Nature Publishing Group},
 	date = {2019},
 	publisher = {Curran Associates, Inc.},
 	pages = {8024{\textendash}8035},
-	url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-performance-deep-learning-library.pdf}
+	url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-performance-deep-learning-library.pdf},
+	doi={https://doi.org/10.48550/arXiv.1912.01703}
 }
 
 @article{martínabadi2015,
@@ -188,14 +196,16 @@ Publisher: Nature Publishing Group},
[…]
