+++ paper/references.bib
@@ -26,7 +26,7 @@
   author={Clink, D. J. and Klinck, Holger},
   journal={arXiv preprint arXiv:1906.02572},
   year={2019},
-  doi={https://doi.org/10.48550/arXiv.1906.02572}
+  doi={10.48550/arXiv.1906.02572}
 }
 
 @article{clinkzenodo2024,
@@ -40,7 +40,7 @@
   year         = 2024,
   publisher    = {Zenodo},
   doi          = {10.5281/zenodo.14213067},
-  url          = {https://doi.org/10.5281/zenodo.14213067},
+  url          = {10.5281/zenodo.14213067},
 }
 
 @article{araya2017warbler,
@@ -114,7 +114,7 @@
 	volume = {11},
 	number = {7},
 	note = {Publisher: Wiley Online Library},
-	doi={https://doi.org/10.1111/2041-210X.13397}
+	doi={10.1111/2041-210X.13397}
 }
 
 @article{kalan2015,
@@ -188,7 +188,7 @@ Publisher: Nature Publishing Group},
 	publisher = {Curran Associates, Inc.},
 	pages = {8024{\textendash}8035},
 	url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-performance-deep-learning-library.pdf},
-	doi={https://doi.org/10.48550/arXiv.1912.01703}
+	doi={10.48550/arXiv.1912.01703}
 }
 
 @article{martínabadi2015,
@@ -197,7 +197,7 @@ Publisher: Nature Publishing Group},
 	year = {2015},
 	date = {2015},
 	url = {https://www.tensorflow.org/},
-	doi ={https://doi.org/10.48550/arXiv.1605.08695}
+	doi ={10.48550/arXiv.1605.08695}
 }
 
 @book{ushey2022,
@@ -213,7 +213,8 @@ Publisher: Nature Publishing Group},
 	author = {Falbel, Daniel},
 	year = {2023},
 	date = {2023},
-	url = {https://CRAN.R-project.org/package=luz}
+	url = {https://CRAN.R-project.org/package=luz},
+	doi={10.32614/CRAN.package.luz}
 }
 
 @book{stevens2020,
@@ -275,7 +276,8 @@ Publisher: Nature Publishing Group},
 	year = {2018},
 	date = {2018},
 	publisher = {Springer},
-	pages = {270{\textendash}279}
+	pages = {270{\textendash}279},
+	doi={10.1007/978-3-030-01424-7_27}
 }
 
 @article{weiss2016,
@@ -287,7 +289,8 @@ Publisher: Nature Publishing Group},
 	pages = {1{\textendash}40},
 	volume = {3},
 	number = {1},
-	note = {Publisher: SpringerOpen}
+	note = {Publisher: SpringerOpen},
+	doi={10.1186/s40537-016-0043-6}
 }
 
 @book{keydana2023,
@@ -299,7 +302,7 @@ Publisher: Nature Publishing Group},
 	doi = {10.1201/9781003275923}
 }
 
-@article{lecun1995,
+@book{lecun1995,
 	title = {Convolutional networks for images, speech, and time series},
 	author = {LeCun, Yann and Bengio, Yoshua and others, },
 	year = {1995},
@@ -319,7 +322,8 @@ Publisher: Nature Publishing Group},
 	journal = {Pattern recognition},
 	pages = {354{\textendash}377},
 	volume = {77},
-	note = {Publisher: Elsevier}
+	note = {Publisher: Elsevier},
+	doi={10.1016/j.patcog.2017.10.013}
 }
 
 @article{krizhevsky2017,
@@ -550,7 +554,8 @@ abstract = {Passive acoustic monitoring – an approach that utilizes autonomous
[…]
