+++ paper/references.bib
@@ -52,7 +52,7 @@
   pages={184--191},
   year={2017},
   publisher={Wiley Online Library},
-  doi={https://dx.doi.org/10.1111/2041-210X.12624}
+  doi={10.1111/2041-210X.12624}
 }
 
 @article{clink2023,
