### 2025-04-29 #7250 @faroit
> I think we are waiting until @steffilazerte gets back on April 25 and hopefully she can let us know regarding the remaining issues. Thanks everyone!

@steffilazerte if you are back, could you please check the post above https://github.com/openjournals/joss-reviews/issues/7250#issuecomment-2802883929 and check if the authors have addressed your comments sufficiently?
