### 2025-05-01 DenaJGibbon/gibbonNetR#5 @steffilazerte
Generally, the documentation update is great! It much clearer to me how everything flows together.

One last suggestion I have (not required for the JOSS review), would be to be even _more_ explicit about the inputs/outputs. For example, in the docs for `get_best_performance()`, under the `performancetables.dir` argument you could modify to

"Path to the directory containing the performance tables. **Output of `train_CNN_binary()` or `train_CNN_multi()`**"

 (by including the function name in \` \` quotes in the docs, pkgdown will automatically link to the docs for that function).
