### 2025-05-01 DenaJGibbon/gibbonNetR#7 @steffilazerte
Great!

Just a couple of extra comments, nothing required for the JOSS review:

The first note is only a problem if you want to submit to CRAN. (Consider setting [R-Universe](https://r-universe.dev/search), if you want something a bit easier for people to install, but don't want to go the CRAN route, but leaving things as is is also perfectly acceptable).

The second note just means that you should add `^Contributing.Rmd$` and `^paper$` to your `.Rbuildignore` file (which you'll probably want to un-gitignore (if possible).
