### 2025-05-01 DenaJGibbon/gibbonNetR#13 @steffilazerte
Complete Example
Hi @DenaJGibbon

I'm just finalizing my review for JOSS, and I'm having a couple of problems with your Complete Example.

- When I run the files, I don't get the same I don't get the same figures as in the pkgdown docs online. Is this to be expected? If so, you may want to mention something to this effect so users aren't confused.

- I also get a different spectrogram

![Image](https://github.com/user-attachments/assets/47fd535f-c5f6-4b7d-bbe0-4b2cdc99e4b8)

- And a different final image

![Image](https://github.com/user-attachments/assets/a32e553c-09c0-4190-a3bf-8ff5145a8752)

If this is just do to stochasticity, then perhaps just add a note to that effect so users aren't confused if they get different results?
