### 2025-05-01 #7250 @steffilazerte
Hi @DenaJGibbon! You've clearly put a lot of work into addressing my comments and suggestions. The package documentation is much better and I have a clear understanding of the relationships among the functions (the flow chart is a great addition). I appreciate the pkgdown site and all the vignettes, I think this will help new users a lot!

I was able to run the examples in the README without problems (as well as the Complete Example 🎉), and you've really improved the best practices. The manuscript reads much more clearly (from the perspective of someone unfamiliar with machine learning) and with the exception of a couple of minor technical tweaks which I think should be addressed, I'm completely satisfied!

I wasn't able to recreate the images in the Complete Example (see https://github.com/DenaJGibbon/gibbonNetR/issues/13), but this may just be due to randomization. So if that's to be expected, perhaps add a note for the user to this effect. And then once the [minor manuscript tweaks](https://github.com/DenaJGibbon/gibbonNetR/issues/8#issuecomment-2845239252) have been addressed (I don't need to review again), I think you're ready to start the post-review ([according](https://github.com/openjournals/joss-reviews/issues/7250#issuecomment-2802883929) to @faroit).

Thanks and congratulations on your hard work!

[@faroit I have checked off "Functional claims", but just to remind you that all I am experienced enough to test is that functions worked without errors and produced output that seemed in line with other examples]
