### 2025-06-06 ba827a7
Update complete example vignette.
