### 2025-06-06 #7250 @DenaJGibbon
I have made the required updates as indicated below. Thank you all for your support and patience on this! It has been a wonderful, positive experience and I will encourage my colleagues to submit to JOSS! Please let me know of other required updates or changes.

**Done**
**Version: [v1.0.1](https://github.com/DenaJGibbon/gibbonNetR/tree/v1.0.1)**
**DOI: https://doi.org/10.5281/zenodo.15608992**
**Done**
**Done** MIT for both.
