### 2025-06-07 #7250 @Kevin-Mattheus-Moerman
- [ ] In the affiliations you need to introduce a space in `Conservation,Universiti` so it reads `Conservation, Universiti`
- [ ] Both `pretrained` and `pre-trained` occur in the text. Please use one form consistently.
