+++ paper/paper.md
@@ -13,9 +13,9 @@ authors:
     affiliation: 2
     orcid: 0000-0001-9658-8600
 affiliations:
- - name: K. Lisa Yang Center for Conservation Bioacoustics,Cornell Lab of Ornithology, Cornell University, Ithaca, New York, United States
+ - name: K. Lisa Yang Center for Conservation Bioacoustics, Cornell Lab of Ornithology, Cornell University, Ithaca, New York, United States
    index: 1
- - name: Institute for Tropical Biology and Conservation,Universiti Malaysia Sabah (UMS), Kota Kinabalu, Sabah, Malaysia
+ - name: Institute for Tropical Biology and Conservation, Universiti Malaysia Sabah (UMS), Kota Kinabalu, Sabah, Malaysia
    index: 2
 
 date: 2025
@@ -40,7 +40,7 @@ Deep learning does not require feature engineering [@stevens2020], as the algori
 
 ## *Transfer learning*
 
-Training deep learning models generally requires a large amount of training data and substantial computing resources. Transfer learning is an approach wherein the architecture of a pretrained CNN (which is generally trained on a very large dataset) is applied to a new classification problem. For example, CNNs trained on the ImageNet dataset of \> 1 million images [@deng2009] such as ResNet have been applied to automated detection/classification of primate and bird species from PAM data [@dufourq2022; @ruan2022]. Generally, very few practitioners train a CNN from scratch, and there are two common approaches for transfer learning. The first option is to use the CNN as a feature extractor, and train only the last classification layer. The second option is known as 'fine-tuning', where instead of initializing a neural network with random weights, the initialization is done using the pre-trained network. Using these pre-trained weights are valuable because the model has already learned useful feature representations [@pytorch_quantized_transfer_learning]. Both approaches require substantially less computing power than training from scratch. The functions in the 'gibbonNetR' package allow users to train models using both types of transfer learning.
+Training deep learning models generally requires a large amount of training data and substantial computing resources. Transfer learning is an approach wherein the architecture of a pre-trained CNN (which is generally trained on a very large dataset) is applied to a new classification problem. For example, CNNs trained on the ImageNet dataset of \> 1 million images [@deng2009] such as ResNet have been applied to automated detection/classification of primate and bird species from PAM data [@dufourq2022; @ruan2022]. Generally, very few practitioners train a CNN from scratch, and there are two common approaches for transfer learning. The first option is to use the CNN as a feature extractor, and train only the last classification layer. The second option is known as 'fine-tuning', where instead of initializing a neural network with random weights, the initialization is done using the pre-trained network. Using these pre-trained weights are valuable because the model has already learned useful feature representations [@pytorch_quantized_transfer_learning]. Both approaches require substantially less computing power than training from scratch. The functions in the 'gibbonNetR' package allow users to train models using both types of transfer learning.
 
 ## State of the field
 
@@ -50,7 +50,7 @@ Until recently, deep learning implementations in R relied on the 'reticulate' pa
 
 # Overview
 
[…]
