### 2025-06-07 d66e49b
Small edits to JOSS paper.
