+++ paper/references.bib
@@ -13,7 +13,7 @@
 
 @article{seewave2008,
     title = {Seewave: a free modular tool for sound analysis and synthesis},
-    author = {{J. Sueur} and {T. Aubin} and {C. Simonis}},
+    author = {Sueur, J. and Aubin, T. and Simonis, C.},
     year = {2008},
     journal = {Bioacoustics},
     volume = {18},
