### 2025-07-08 #7250 @faroit
@DenaJGibbon better late than never: thanks again for submitting your work to JOSS. I am sorry that I couldn't help that much on technical guidance, as R isn't something I regularly use but I think the software and the paper has greatly improved during this process. Thanks to @steffilazerte @Desjonqu for your editing work that made this possible!
